@@ -1,72 +1,78 @@
 {
-  "name": "forpdi-web",
-  "version": "2.0.0-beta1",
-  "description": "ForPDI web client.",
-  "scripts": {
-    "dev": "webpack-dev-server --colors --compress --progress --port 8081 --inline -w",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack --colors --progress --config webpack.production.config.js -p",
-    "build-dev": "webpack --progress --colors -d"
-  },
-  "author": "Renato R. R. de Oliveira <renatorro@comp.ufla.br>, Rodrigo Freitas <rodrigotkd@gmail.com>",
-  "devDependencies": {
-    "autoprefixer": "^9.4.9",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.0.0",
-    "babel-plugin-transform-async-to-generator": "^6.24.1",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-plugin-webpack-aliases": "^1.1.3",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "css-loader": "^1.0.1",
-    "exports-loader": "^0.7.0",
-    "file-loader": "^2.0.0",
-    "imports-loader": "^0.8.0",
-    "mdi": "^1.5.54",
-    "node-sass": "^4.11.0",
-    "postcss-loader": "^3.0.0",
-    "precss": "^3.1.2",
-    "sass-loader": "^7.1.0",
-    "style-loader": "^0.23.1",
-    "url-loader": "^1.1.2",
-    "webpack": "^4.29.6",
-    "webpack-cli": "^3.2.3"
-  },
-  "dependencies": {
-    "backbone": "^1.4.0",
-    "blueimp-file-upload": "^9.28.0",
-    "bootstrap": "^3.4.1",
-    "flux": "^3.1.3",
-    "history": "^4.7.2",
-    "jquery": "3.3.1",
-    "marked": "^0.5.2",
-    "moment": "^2.24.0",
-    "numeral": "^1.5.3",
-    "quill": "^1.3.6",
-    "react": "15.x.x",
-    "react-bootstrap": "^0.32.4",
-    "react-date-picker": "^6.13.0",
-    "react-datepicker": "^0.29.0",
-    "react-dom": "15.x.x",
-    "react-google-charts": "^1.6.6",
-    "react-maskedinput": "^4.0.1",
-    "react-onclickoutside": "^5.11.1",
-    "react-progressbar": "^15.4.1",
-    "react-quill": "^1.3.3",
-    "react-router": "^2.4.0",
-    "react-toastr": "^2.8.1",
-    "react-tooltip": "^3.9.2",
-    "string": "^3.3.3",
-    "toastr": "^2.1.4",
-    "underscore": "1.9.1",
-    "webpack-dev-server": "^3.2.1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/forpdi/forpdi"
-  },
-  "license": "Apache-2.0",
-  "readme": "../../../../README.md"
+	"name": "forpdi-web",
+	"version": "2.0.0-beta1",
+	"description": "ForPDI web client.",
+	"scripts": {
+		"dev": "webpack-dev-server --colors --compress --progress --port 8081 --inline -w",
+		"test": "echo \"Error: no test specified\" && exit 1",
+		"build": "webpack --colors --progress --config webpack.production.config.js -p",
+		"build:docker": "./node_modules/.bin/webpack --config webpack.production.config.js -p",
+		"build:dev": "webpack --progress --colors -d"
+	},
+	"author": "Renato R. R. de Oliveira <renatorro@comp.ufla.br>, Rodrigo Freitas <rodrigotkd@gmail.com>",
+	"devDependencies": {
+		"autoprefixer": "^9.6.1",
+		"babel-core": "^6.26.3",
+		"babel-loader": "^7.0.0",
+		"babel-plugin-transform-async-to-generator": "^6.24.1",
+		"babel-plugin-transform-class-properties": "^6.24.1",
+		"babel-plugin-transform-object-rest-spread": "^6.26.0",
+		"babel-plugin-transform-runtime": "^6.23.0",
+		"babel-plugin-webpack-aliases": "^1.1.3",
+		"babel-preset-env": "^1.7.0",
+		"babel-preset-react": "^6.24.1",
+		"css-loader": "^1.0.1",
+		"exports-loader": "^0.7.0",
+		"file-loader": "^2.0.0",
+		"imports-loader": "^0.8.0",
+		"node-sass": "^4.12.0",
+		"postcss-loader": "^3.0.0",
+		"precss": "^4.0.0",
+		"sass-loader": "^7.1.0",
+		"style-loader": "^0.23.1",
+		"url-loader": "^1.1.2",
+		"webpack": "^4.36.1",
+		"webpack-cli": "^3.2.6",
+		"webpack-dev-server": "^3.7.2"
+	},
+	"dependencies": {
+		"backbone": "^1.4.0",
+		"blueimp-file-upload": "^9.28.0",
+		"bootstrap": "^3.4.1",
+		"flux": "^3.1.3",
+		"history": "^4.7.2",
+		"jquery": "3.3.1",
+		"marked": "^0.5.2",
+		"mdi": "^1.5.54",
+		"moment": "^2.24.0",
+		"numeral": "^1.5.3",
+		"prop-types": "^15.7.2",
+		"quill": "^1.3.6",
+		"react": "15.x.x",
+		"react-bootstrap": "^0.32.4",
+		"react-date-picker": "^6.13.0",
+		"react-datepicker": "^0.29.0",
+		"react-dom": "15.x.x",
+		"react-google-charts": "^1.6.6",
+		"react-helmet": "^5.2.0",
+		"react-maskedinput": "^4.0.1",
+		"react-multiselect-checkboxes": "^0.1.1",
+		"react-onclickoutside": "^5.11.1",
+		"react-progressbar": "^15.4.1",
+		"react-quill": "^1.3.3",
+		"react-router": "^2.4.0",
+		"react-select": "^2.4.2",
+		"react-table": "^6.9.2",
+		"react-toastr": "^2.8.1",
+		"react-tooltip": "^3.9.2",
+		"string": "^3.3.3",
+		"toastr": "^2.1.4",
+		"underscore": "1.9.1"
+	},
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/forpdi/forpdi"
+	},
+	"license": "Apache-2.0",
+	"readme": "../../../../README.md"
 }
