@@ -20,7 +20,11 @@ public enum NotificationType implements Serializable{
 	ACTION_PLAN_CLOSE_TO_MATURITY(15, "http://cloud.progolden.com.br/file/8343", false),
 	SEND_MESSAGE(16, "http://cloud.progolden.com.br/file/8498", false),
 	DATE_ATTRIBUTE_UPDATED(17, " http://cloud.progolden.com.br/file/8707", false),
-	GOAL_OPENED(18, "http://cloud.progolden.com.br/file/8337", false);
+	GOAL_OPENED(18, "http://cloud.progolden.com.br/file/8337", false),
+	FORRISCO_PROCESS_CREATED(19, "", true),
+	FORRISCO_RISK_CLOSE_TO_MATURITY(20, "http://cloud.progolden.com.br/file/8343",false),
+	FORRISCO_USER_LINKED_TO_RISK(20, "http://cloud.progolden.com.br/file/8336", false);
+
 	
 	private Integer value;
 	private String imageUrl;
