@@ -2,13 +2,12 @@ package org.forpdi.planning.attribute.types;
 
 import org.forpdi.planning.attribute.AttributeType;
 import org.forpdi.planning.attribute.AttributeTypeWrapper;
-import org.forpdi.planning.attribute.types.ActionPlanField.Wrapper;
 
 public class StrategicObjective extends AttributeType {
 	private AttributeTypeWrapper wrapper = new Wrapper();
 
 	public final String WIDGET_NAME = "StrategicObjectivesField";
-	
+
 	@Override
 	public String getWidget() {
 		return WIDGET_NAME;
@@ -20,21 +19,21 @@ public class StrategicObjective extends AttributeType {
 	}
 
 	@Override
-	public String getDisplayName() {		
+	public String getDisplayName() {
 		return "Objetivos estratégicos";
 	}
-	
+
 	public static class Wrapper extends TextField.Wrapper {
-		
+
 		@Override
 		public String fromDatabase(String databaseValue) {
 			return null;
 		}
-		
+
 		@Override
 		public String toDatabase(String viewValue) {
 			return null;
 		}
-		
+
 	}
 }
