@@ -1,39 +1,29 @@
 package org.forpdi.planning.document;
 
-import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
-import java.io.FileReader;
-import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.sql.SQLException;
 import java.text.DecimalFormat;
 import java.text.NumberFormat;
 import java.text.SimpleDateFormat;
-import java.util.AbstractMap.SimpleEntry;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Collections;
 import java.util.Date;
 import java.util.HashMap;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.Queue;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 
 import org.forpdi.core.company.CompanyDomain;
 import org.forpdi.core.event.Current;
-import org.forpdi.core.properties.SystemConfigs;
 import org.forpdi.core.user.User;
 import org.forpdi.core.user.UserBS;
 import org.forpdi.planning.attribute.AggregateIndicator;
@@ -53,7 +43,6 @@ import org.forpdi.planning.attribute.types.SelectField;
 import org.forpdi.planning.attribute.types.SelectPlan;
 import org.forpdi.planning.attribute.types.StrategicObjective;
 import org.forpdi.planning.attribute.types.TableField;
-import org.forpdi.planning.attribute.types.TextArea;
 import org.forpdi.planning.attribute.types.enums.FormatValue;
 import org.forpdi.planning.fields.FieldsBS;
 import org.forpdi.planning.fields.OptionsField;
@@ -78,7 +67,6 @@ import org.hibernate.criterion.Order;
 import org.hibernate.criterion.Projections;
 import org.hibernate.criterion.Restrictions;
 
-import com.itextpdf.text.Chunk;
 import com.itextpdf.text.DocumentException;
 import com.itextpdf.text.Element;
 import com.itextpdf.text.Font;
@@ -88,22 +76,15 @@ import com.itextpdf.text.PageSize;
 import com.itextpdf.text.Paragraph;
 import com.itextpdf.text.Phrase;
 import com.itextpdf.text.Rectangle;
-import com.itextpdf.text.html.simpleparser.HTMLWorker;
-import com.itextpdf.text.html.simpleparser.StyleSheet;
 import com.itextpdf.text.pdf.BaseFont;
 import com.itextpdf.text.pdf.CMYKColor;
 import com.itextpdf.text.pdf.ColumnText;
-import com.itextpdf.text.pdf.PdfAction;
 import com.itextpdf.text.pdf.PdfContentByte;
-import com.itextpdf.text.pdf.PdfCopy;
-import com.itextpdf.text.pdf.PdfImportedPage;
 import com.itextpdf.text.pdf.PdfPCell;
 import com.itextpdf.text.pdf.PdfPTable;
-import com.itextpdf.text.pdf.PdfPageEventHelper;
 import com.itextpdf.text.pdf.PdfReader;
 import com.itextpdf.text.pdf.PdfStamper;
 import com.itextpdf.text.pdf.PdfWriter;
-import com.itextpdf.text.pdf.draw.DottedLineSeparator;
 
 import br.com.caelum.vraptor.boilerplate.HibernateBusiness;
 import br.com.caelum.vraptor.boilerplate.bean.PaginatedList;
@@ -846,866 +827,6 @@ public class DocumentBS extends HibernateBusiness {
 				}
 			}
 		}
-	}
-
-	/**
-	 * Exporta documento PDF
-	 * 
-	 * @param author
-	 *            Autor do documento.
-	 * @param title
-	 *            Título do documento.
-	 * @param lista
-	 *            Lista de seções a serem exportadas.
-	 * @return InputStream Fluxo de entrada do arquivo PDF.
-	 * @throws IOException
-	 * @throws DocumentException
-	 * @throws SQLException
-	 * @throws ClassNotFoundException
-	 */
-	public InputStream exportDocument(String author, String title, String lista)
-			throws IOException, DocumentException, SQLException, ClassNotFoundException {
-
-		class TOCEvent extends PdfPageEventHelper {
-
-			protected int counter = 0;
-			protected List<SimpleEntry<String, SimpleEntry<String, Integer>>> toc = new ArrayList<>();
-
-			String lastText = "";
-
-			@Override
-			public void onGenericTag(PdfWriter writer, com.itextpdf.text.Document document, Rectangle rect,
-					String text) {
-				if (text != lastText) {
-					String name = "dest" + (counter++);
-					int page = writer.getPageNumber();
-					// LOGGER.info(text);
-					toc.add(new SimpleEntry<String, SimpleEntry<String, Integer>>(text,
-							new SimpleEntry<String, Integer>(name, page)));
-				}
-				lastText = text;
-				// writer.addNamedDestination(name, page, new
-				// PdfDestination(PdfDestination.FITH, rect.getTop()));
-
-			}
-
-			public List<SimpleEntry<String, SimpleEntry<String, Integer>>> getTOC() {
-				return toc;
-			}
-
-		}
-
-		com.itextpdf.text.Document document = new com.itextpdf.text.Document();
-		com.itextpdf.text.Document coverDocument = new com.itextpdf.text.Document();
-		com.itextpdf.text.Document preTextDocument = new com.itextpdf.text.Document();
-		com.itextpdf.text.Document summaryDocument = new com.itextpdf.text.Document();
-
-		File outputDir;
-		final String storagePath = SystemConfigs.getConfig("store.pdfs");
-		if (storagePath == null || storagePath.equals("") || storagePath.equals("${store.pdfs}")) {
-			outputDir = File.createTempFile("fpdi-document-export", ".pdf").getParentFile();
-		} else {
-			outputDir = new File(storagePath);
-			if (!outputDir.exists()) {
-				if (!outputDir.mkdirs()) {
-					throw new RuntimeException("Failed to create storage directory.");
-				}
-			} else if (!outputDir.isDirectory()) {
-				throw new RuntimeException("The configured storage path is not a directory.");
-			}
-		}
-		
-		final String prefix = String.format("fpdi-doc-export-%d", System.currentTimeMillis());
-
-		File coverPdfFile = new File(outputDir, String.format("%s-cover.pdf", prefix));
-		File preTextPdfFile = new File(outputDir, String.format("%s-pre-text.pdf", prefix));
-		File summaryPdfFile = new File(outputDir, String.format("%s-summary.pdf", prefix));
-		File finalSummaryPdfFile = new File(outputDir, String.format("%s-final-summary.pdf", prefix));
-		File contentFile = new File(outputDir, String.format("%s-content.pdf", prefix));
-		File destinationFile = new File(outputDir, String.format("%s-mounted.pdf", prefix));
-		File finalPdfFile = new File(outputDir, String.format("%s-final.pdf", prefix));
-
-		PdfWriter writer = PdfWriter.getInstance(document, new FileOutputStream(contentFile));
-		PdfWriter coverWriter = PdfWriter.getInstance(coverDocument, new FileOutputStream(coverPdfFile));
-		PdfWriter.getInstance(preTextDocument, new FileOutputStream(preTextPdfFile));
-		PdfWriter.getInstance(summaryDocument, new FileOutputStream(summaryPdfFile));
-
-		TOCEvent event = new TOCEvent();
-		writer.setPageEvent(event);
-
-		// DEFINIÇÕES DE FONTE, MARGENS, ESPAÇAMENTO E CORES
-		Font texto = FontFactory.getFont(FontFactory.TIMES, 12.0f);
-
-		Font titulo = FontFactory.getFont(FontFactory.TIMES_BOLD, 12.0f);
-		Font tituloCapa = FontFactory.getFont(FontFactory.TIMES_BOLD, 14.0f);
-		// Cor cinza - cabeçalho das tabelas
-		CMYKColor headerBgColor = new CMYKColor(55, 45, 42, 7);
-
-		// 0,8 cm acima e abaixo
-		float paragraphSpacing = 22.6772f;
-		// Parágrafo com 1,25 cm na primeira linha
-		float firstLineIndent = 35.43307f;
-		// 1,5 entrelinhas
-		float interLineSpacing = texto.getCalculatedLeading(1.5f);
-		// Formato A4 do documento
-		coverDocument.setPageSize(PageSize.A4);
-		// Margens Superior e esquerda: 3 cm Inferior e direita: 2 cm
-		coverDocument.setMargins(85.0394f, 56.6929f, 85.0394f, 56.6929f);
-
-		coverDocument.open();
-
-		// CABEÇALHO
-		String imageUrl = domain.getCompany().getLogo();
-		// LOGGER.info("|"+imageUrl+"|");
-		if (!imageUrl.trim().isEmpty()) {
-			Image image = Image.getInstance(new URL(imageUrl));
-			// image.scaleAbsolute(150f, 150f);
-			float scaler = ((coverDocument.getPageSize().getWidth() - coverDocument.leftMargin()
-					- coverDocument.rightMargin()) / image.getWidth()) * 100;
-			image.scalePercent(scaler * 0.4f);
-			image.setAlignment(Element.ALIGN_CENTER);
-			coverDocument.add(image);
-		}
-
-		Paragraph TITULO = new Paragraph(title, tituloCapa);
-		// Paragraph AUTHOR = new Paragraph(author, texto);
-		TITULO.setAlignment(Element.ALIGN_CENTER);
-		TITULO.setSpacingBefore(paragraphSpacing * 8);
-
-		// AUTHOR.setAlignment(Element.ALIGN_CENTER);
-
-		coverDocument.add(TITULO);
-		// document.add(AUTHOR);
-		// document.add(YEAR);
-		Phrase localizationPhrase = new Phrase(this.domain.getCompany().getLocalization(), titulo);
-		// Phrase footer = new Phrase(String.valueOf(cal.get(Calendar.YEAR)),
-		// titulo);
-		PdfContentByte cb = coverWriter.getDirectContent();
-		ColumnText.showTextAligned(cb, Element.ALIGN_CENTER, localizationPhrase,
-				(coverDocument.right() - coverDocument.left()) / 2 + coverDocument.leftMargin(),
-				coverDocument.bottom() + 30, 0);
-		coverDocument.newPage();
-
-		// FOLHA DE ROSTO
-		Paragraph COMPANY = new Paragraph(domain.getCompany().getName(), tituloCapa);
-		COMPANY.setAlignment(Element.ALIGN_CENTER);
-		COMPANY.setSpacingBefore(paragraphSpacing);
-		coverDocument.add(COMPANY);
-
-		coverDocument.add(TITULO);
-
-		Calendar cal = Calendar.getInstance();
-
-		Phrase periodPhrase = new Phrase(
-				String.valueOf(cal.get(Calendar.MONTH) + 1) + "/" + String.valueOf(cal.get(Calendar.YEAR)), titulo);
-		// footerRosto.add(periodPhrase);
-		// footerRosto.add(footer);
-		ColumnText.showTextAligned(cb, Element.ALIGN_CENTER, periodPhrase,
-				(coverDocument.right() - coverDocument.left()) / 2 + coverDocument.leftMargin(),
-				coverDocument.bottom() + 15, 0);
-		ColumnText.showTextAligned(cb, Element.ALIGN_CENTER, localizationPhrase,
-				(coverDocument.right() - coverDocument.left()) / 2 + coverDocument.leftMargin(),
-				coverDocument.bottom() + 30, 0);
-		coverDocument.newPage();
-		coverDocument.close();
-
-		String[] sections = lista.split(",");
-		int secIndex = 0, subSecIndex = 0;
-
-		boolean lastAttWasPlan = false;
-
-		document.setPageSize(PageSize.A4);
-		// Margens Superior e esquerda: 3 cm Inferior e direita: 2 cm
-		document.setMargins(85.0394f, 56.6929f, 85.0394f, 56.6929f);
-		document.open();
-
-		preTextDocument.setPageSize(PageSize.A4);
-		// Margens Superior e esquerda: 3 cm Inferior e direita: 2 cm
-		preTextDocument.setMargins(85.0394f, 56.6929f, 85.0394f, 56.6929f);
-		preTextDocument.open();
-
-		boolean havePreText = false;
-		boolean haveContent = false;
-
-		for (int i = 0; i < sections.length; i++) {
-			DocumentSection ds = this.retrieveSectionById(Long.parseLong(sections[i]));
-			ds.setDocumentAttributes(this.listAttributesBySection(ds, ds.getDocument().getPlan().getId()));
-
-			subSecIndex = 0;
-			String secName = ds.getName();
-
-			if (ds.isPreTextSection()) { // SEÇÕES PRÉ TEXTUAIS
-
-				for (DocumentAttribute a : ds.getDocumentAttributes()) {
-					if (a.getType().equals(TableField.class.getCanonicalName())) {
-						havePreText = true;
-						TableFields tf = fieldsBS.tableFieldsByAttribute(a.getId(), true, false);
-						List<TableStructure> tabStructList = fieldsBS.listTableStructureByFields(tf);
-						List<TableInstance> tabInstList = fieldsBS.listTableInstanceByFields(tf,false);
-						if (!tabInstList.isEmpty()) {
-							// String attName = a.getName();
-							// if (!attName.equals(secName)) {
-
-							Chunk c = new Chunk(secName, titulo);
-							c.setGenericTag(secName);
-
-							Paragraph attTitle = new Paragraph(c);
-							attTitle.setAlignment(Element.ALIGN_CENTER);
-							attTitle.setLeading(interLineSpacing);
-							attTitle.setSpacingAfter(paragraphSpacing);
-							attTitle.setSpacingBefore(paragraphSpacing);
-							preTextDocument.add(attTitle);
-							// }
-							PdfPTable table;
-							if (tabStructList
-									.size() == 4) /**
-													 * SOLUÇÃO DE CONTORNO PARA
-													 * TRATAR O CASO ESPECIAL DO
-													 * HISTÓRICO DE VERSÕES QUE
-													 * É UMA SEÇÃO PRÉ-TEXTUAL
-													 * MAS DEVE EXIBIR BORDA E
-													 * CABEÇALHO, NECESSITA DE
-													 * MELHORIA
-													 */
-								table = returnPdfPTable(tabStructList, tabInstList, false, false);
-							else
-								table = returnPdfPTable(tabStructList, tabInstList, true, true);
-							preTextDocument.add(table);
-							preTextDocument.newPage();
-							lastAttWasPlan = false;
-						}
-					}
-				}
-
-			} else {
-				haveContent = true;
-				// SEÇÕES NUMERADAS
-				/*
-				 * if (lastSecWasPreText) { lastSecWasPreText = false;
-				 * 
-				 * Paragraph secTitle = new Paragraph("Sumário", titulo);
-				 * secTitle.setLeading(interLineSpacing);
-				 * secTitle.setSpacingAfter(paragraphSpacing);
-				 * secTitle.setSpacingBefore(paragraphSpacing);
-				 * secTitle.setAlignment(Element.ALIGN_CENTER);
-				 * document.add(secTitle); int summaryIndex = 0; int
-				 * summarySubSecIndex = 0; for (String secaoId : sections) {
-				 * summarySection =
-				 * this.retrieveSectionById(Long.parseLong(secaoId)); if
-				 * (!summarySection.isPreTextSection()) {
-				 * 
-				 * summaryIndex++; secTitle = new Paragraph(summaryIndex + ". "
-				 * + summarySection.getName(), titulo);
-				 * secTitle.setLeading(interLineSpacing);
-				 * document.add(secTitle);
-				 * 
-				 * List<DocumentSection> dsList =
-				 * this.listSectionsSons(summarySection);
-				 * this.setSectionsFilled(dsList,
-				 * summarySection.getDocument().getPlan().getId());
-				 * summarySubSecIndex = 0; for (DocumentSection sec : dsList) {
-				 * 
-				 * if (sec.isFilled()) { summarySubSecIndex++; secTitle = new
-				 * Paragraph( summaryIndex + "." + summarySubSecIndex + ". " +
-				 * sec.getName(), texto); secTitle.setLeading(interLineSpacing);
-				 * secTitle.setFirstLineIndent(firstLineIndent / 2);
-				 * document.add(secTitle); } }
-				 * 
-				 * } } document.newPage(); }
-				 */
-
-				secIndex++;
-
-				// LOGGER.info(ds.getId() + ". " + ds.getName() + " - Size:
-				// " +
-				// ds.getDocumentAttributes().size());
-
-				boolean secTitlePrinted = false;
-
-				if (ds.getDocumentAttributes().size() == 0) {
-					if (lastAttWasPlan) {
-						document.setPageSize(PageSize.A4);
-						document.newPage();
-					}
-					Chunk c = new Chunk(secIndex + ". " + secName, titulo);
-					c.setGenericTag(secIndex + ". " + secName);
-					Paragraph secTitle = new Paragraph(c);
-					secTitle.setLeading(interLineSpacing);
-					secTitle.setSpacingAfter(paragraphSpacing);
-					secTitle.setSpacingBefore(paragraphSpacing);
-					document.add(secTitle);
-					secTitlePrinted = true;
-					lastAttWasPlan = false;
-				}
-
-				for (DocumentAttribute a : ds.getDocumentAttributes()) {
-					if (a.getType().equals(TextArea.class.getCanonicalName())) {
-						if (a.getValue() != null && !a.getValue().equals("")) {
-							if (lastAttWasPlan) {
-								document.setPageSize(PageSize.A4);
-								document.newPage();
-							}
-							if (!secTitlePrinted) {
-								Chunk c = new Chunk(secIndex + ". " + secName, titulo);
-								c.setGenericTag(secIndex + ". " + secName);
-								Paragraph secTitle = new Paragraph(c);
-								secTitle.setLeading(interLineSpacing);
-								secTitle.setSpacingAfter(paragraphSpacing);
-								secTitle.setSpacingBefore(paragraphSpacing);
-								document.add(secTitle);
-								secTitlePrinted = true;
-							}
-							String attName = a.getName();
-							if (!attName.equals(secName)) {
-								Paragraph attTitle = new Paragraph(attName, titulo);
-								attTitle.setLeading(interLineSpacing);
-								attTitle.setSpacingAfter(paragraphSpacing);
-								attTitle.setSpacingBefore(paragraphSpacing);
-								document.add(attTitle);
-							}
-							// HTMLWorker htmlWorker = new
-							// HTMLWorker(document);
-							Map<String, String> pc2 = new HashMap<String, String>();
-							pc2.put("line-height", "115%");
-							pc2.put("margin-bottom", "6.0pt");
-							pc2.put("text-align", "center");
-
-							HashMap<String, String> spanc1 = new HashMap<String, String>();
-							spanc1.put("text-justify", "inter-word");
-
-							StyleSheet styles = new StyleSheet();
-							styles.loadTagStyle("p", "text-indent", "1.25cm");
-
-							String str = "<html>" + "<head>" + "</head><body style=\"text-indent: 1.25cm; \">"
-									+ "<p style=\"text-indent: 1.25cm; \">";
-							Queue<String> allMatches = new LinkedList<>();
-							String value = a.getValue();
-							if (a.getValue().contains("<img")) {
-								Matcher m = Pattern.compile("<img [^>]*>").matcher(a.getValue());
-								while (m.find()) {
-									String match = m.group();
-									allMatches.add(match);
-									value = value.replace(match, "<p>||IMAGE||</p>");
-								}
-							}
-							str += value + "</p></body></html>";
-
-							File htmlFile = new File(outputDir, String.format("%s-1.html", prefix));
-							FileWriter fw = new FileWriter(htmlFile, true);
-							BufferedWriter conexao = new BufferedWriter(fw);
-							conexao.write(str);
-							conexao.newLine();
-							conexao.close();
-							// LOGGER.info(htmlFile.getPath());
-
-							List<Element> p = HTMLWorker.parseToList(new FileReader(htmlFile.getPath()), styles);
-							for (int k = 0; k < p.size(); ++k) {
-								if (p.get(k) instanceof Paragraph) {
-									Paragraph att = (Paragraph) p.get(k);
-									// LOGGER.info("------->"+att.getContent());
-									if (att.getContent().contains("||IMAGE||")) {
-										String img = allMatches.poll();
-										if (img != null) {
-											// LOGGER.info("IMG------->"+img);
-											Image image = Image.getInstance(
-													new URL(img.replaceAll("<img src=\"", "").replaceAll("\">", "")));
-											float scaler = ((document.getPageSize().getWidth() - document.leftMargin()
-													- document.rightMargin()) / image.getWidth()) * 100;
-											image.scalePercent(scaler * 0.4f);
-											image.setAlignment(Element.ALIGN_CENTER);
-											document.add(image);
-										}
-									} else {
-										att.setFirstLineIndent(firstLineIndent);
-										document.add(att);
-									}
-								} else if (p.get(k).getClass().getName().equals("com.itextpdf.text.List")) {
-									com.itextpdf.text.List att = (com.itextpdf.text.List) p.get(k);
-									att.setIndentationLeft(firstLineIndent);
-									document.add(att);
-								}
-							}
-							lastAttWasPlan = false;
-							htmlFile.delete();
-						}
-					} else if (a.getType().equals(TableField.class.getCanonicalName())) {
-
-						TableFields tf = fieldsBS.tableFieldsByAttribute(a.getId(), true, false);
-						List<TableStructure> tabStructList = fieldsBS.listTableStructureByFields(tf);
-						List<TableInstance> tabInstList = fieldsBS.listTableInstanceByFields(tf,false);
-						if (!tabInstList.isEmpty()) {
-							if (lastAttWasPlan) {
-								document.setPageSize(PageSize.A4);
-								document.newPage();
-							}
-							if (!secTitlePrinted) {
-								Chunk c = new Chunk(secIndex + ". " + secName, titulo);
-								c.setGenericTag(secIndex + ". " + secName);
-								Paragraph secTitle = new Paragraph(c);
-								secTitle.setLeading(interLineSpacing);
-								secTitle.setSpacingAfter(paragraphSpacing);
-								secTitle.setSpacingBefore(paragraphSpacing);
-								document.add(secTitle);
-								secTitlePrinted = true;
-							}
-							String attName = a.getName();
-							if (!attName.equals(secName)) {
-								Paragraph attTitle = new Paragraph(attName, texto);
-								attTitle.setAlignment(Element.ALIGN_CENTER);
-								attTitle.setLeading(interLineSpacing);
-								attTitle.setSpacingAfter(paragraphSpacing);
-								attTitle.setSpacingBefore(paragraphSpacing);
-								document.add(attTitle);
-							}
-							PdfPTable table = returnPdfPTable(tabStructList, tabInstList, false, false);
-							document.add(table);
-							lastAttWasPlan = false;
-						}
-
-					} else if (a.getType().equals(ScheduleField.class.getCanonicalName())) {
-						List<ScheduleInstance> schInstList = this.fieldsBS
-								.retrieveScheduleInstanceByAttribute(a.getId(), true);
-						if (!schInstList.isEmpty()) {
-							if (lastAttWasPlan) {
-								document.setPageSize(PageSize.A4);
-								document.newPage();
-							}
-							if (!secTitlePrinted) {
-								Chunk c = new Chunk(secIndex + ". " + secName, titulo);
-								c.setGenericTag(secIndex + ". " + secName);
-								Paragraph secTitle = new Paragraph(c);
-								secTitle.setLeading(interLineSpacing);
-								secTitle.setSpacingAfter(paragraphSpacing);
-								secTitle.setSpacingBefore(paragraphSpacing);
-								document.add(secTitle);
-								secTitlePrinted = true;
-							}
-							String attName = a.getName();
-							if (!attName.equals(secName)) {
-								Paragraph attTitle = new Paragraph(attName, texto);
-								attTitle.setAlignment(Element.ALIGN_CENTER);
-								attTitle.setLeading(interLineSpacing);
-								attTitle.setSpacingAfter(paragraphSpacing);
-								attTitle.setSpacingBefore(paragraphSpacing);
-								document.add(attTitle);
-							}
-							PdfPTable table = new PdfPTable(3);
-							table.getDefaultCell();
-							PdfPCell c = new PdfPCell(new Paragraph("Atividade", texto));
-							c.setHorizontalAlignment(Element.ALIGN_CENTER);
-							c.setBackgroundColor(headerBgColor);
-							table.addCell(c);
-							c = new PdfPCell(new Paragraph("Início", texto));
-							c.setHorizontalAlignment(Element.ALIGN_CENTER);
-							c.setBackgroundColor(headerBgColor);
-
-							table.addCell(c);
-							c = new PdfPCell(new Paragraph("Fim", texto));
-							c.setHorizontalAlignment(Element.ALIGN_CENTER);
-							c.setBackgroundColor(headerBgColor);
-							table.addCell(c);
-
-							SimpleDateFormat sdf = new SimpleDateFormat("dd/MM/yyyy");
-							for (ScheduleInstance sch : schInstList) {
-								table.addCell(new Paragraph(sch.getDescription(), texto));
-								table.addCell(new Paragraph(sdf.format(sch.getBegin()), texto));
-								table.addCell(new Paragraph(sdf.format(sch.getEnd()), texto));
-							}
-							document.add(table);
-							lastAttWasPlan = false;
-						}
-					} else if (a.getType().equals(SelectPlan.class.getCanonicalName())) {
-						if (a.getValue() != null) {
-							Plan plan = planBS.retrieveById(Long.parseLong(a.getValue()));
-							List<PdfPTable> planTableList = this.generatePDFplanTable(plan);
-							boolean first = true;
-							// LOGGER.info("2 - " + plan.getName());
-							// LOGGER.info(secName + " " + secTitlePrinted);
-							for (PdfPTable planTable : planTableList) {
-								if (!lastAttWasPlan) {
-									document.setPageSize(PageSize.A4.rotate());
-								}
-								document.newPage();
-								if (first) {
-
-									if (!secTitlePrinted) {
-										Chunk c = new Chunk(secIndex + ". " + secName, titulo);
-										c.setGenericTag(secIndex + ". " + secName);
-										Paragraph secTitle = new Paragraph(c);
-										document.add(secTitle);
-										secTitlePrinted = true;
-									}
-									Paragraph attTitle = new Paragraph(plan.getName(), texto);
-									attTitle.setAlignment(Element.ALIGN_CENTER);
-									document.add(attTitle);
-									first = false;
-									lastAttWasPlan = true;
-								}
-								document.add(planTable);
-							}
-							// document.setPageSize(PageSize.A4);
-							// document.newPage();
-						}
-					}
-				}
-
-				List<DocumentSection> dsList = this.listSectionsSons(ds);
-				this.setSectionsFilled(dsList, ds.getDocument().getPlan().getId());
-
-				// LOGGER.info("dsList.size: " + dsList.size());
-
-				for (DocumentSection d : dsList) {
-					if (d.isFilled()) {
-						// LOGGER.info("filled: " + d.getName());
-						String subSecName = d.getName();
-						boolean subSecTitlePrinted = false;
-						// if (!secName.equals(subSecName)) {
-						/*
-						 * subSecIndex++; Paragraph subSecTitle = new
-						 * Paragraph(secIndex + "." + subSecIndex + ". " +
-						 * subSecName, titulo);
-						 * subSecTitle.setLeading(interLineSpacing);
-						 * subSecTitle.setSpacingAfter(paragraphSpacing);
-						 * subSecTitle.setSpacingBefore(paragraphSpacing);
-						 * document.add(subSecTitle); subSecTitlePrinted = true;
-						 */
-						// }
-
-						List<DocumentAttribute> attList = this.listAttributesBySection(d,
-								d.getDocument().getPlan().getId());
-						// LOGGER.info("attList.size: " + attList.size());
-						for (DocumentAttribute a : attList) {
-							if (a.getType().equals(TextArea.class.getCanonicalName())) {
-								if (a.getValue() != null && !a.getValue().equals("")) {
-									String attName = a.getName();
-									if (lastAttWasPlan) {
-										document.setPageSize(PageSize.A4);
-										document.newPage();
-									}
-									if (!subSecTitlePrinted) {
-										subSecIndex++;
-										Chunk c = new Chunk(secIndex + "." + subSecIndex + ". " + subSecName, titulo);
-										c.setGenericTag(secIndex + "." + subSecIndex + ". " + subSecName);
-										Paragraph subSecTitle = new Paragraph(c);
-										subSecTitle.setLeading(interLineSpacing);
-										subSecTitle.setSpacingAfter(paragraphSpacing);
-										subSecTitle.setSpacingBefore(paragraphSpacing);
-										document.add(subSecTitle);
-										subSecTitlePrinted = true;
-									}
-									if (!attName.equals(subSecName)) {
-										Paragraph attTitle = new Paragraph(attName, titulo);
-										attTitle.setLeading(interLineSpacing);
-										attTitle.setSpacingAfter(paragraphSpacing);
-										attTitle.setSpacingBefore(paragraphSpacing);
-										document.add(attTitle);
-									}
-									// HTMLWorker htmlWorker = new
-									// HTMLWorker(document);
-									Map<String, String> pc2 = new HashMap<String, String>();
-									pc2.put("line-height", "115%");
-									pc2.put("margin-bottom", "6.0pt");
-									pc2.put("text-align", "center");
-
-									HashMap<String, String> spanc1 = new HashMap<String, String>();
-									spanc1.put("text-justify", "inter-word");
-
-									StyleSheet styles = new StyleSheet();
-									styles.loadTagStyle("p", "text-indent", "1.25cm");
-
-									String str = "<html>" + "<head>"
-											+ "</head><body style=\"text-indent: 1.25cm; font-size: 12pt !important;\">"
-											+ "<p style=\"text-indent: 1.25cm; font-size: 12pt !important;\">";
-									Queue<String> allMatches = new LinkedList<>();
-									String value = a.getValue();
-									if (a.getValue().contains("<img")) {
-										Matcher m = Pattern.compile("<img [^>]*>").matcher(a.getValue());
-										while (m.find()) {
-											String match = m.group();
-											allMatches.add(match);
-											value = value.replace(match, "<p>||IMAGE||</p>");
-										}
-									}
-									str += value + "</p></body></html>";
-
-									File htmlFile = new File(outputDir, String.format("%s-2.html", prefix));
-									FileWriter fw = new FileWriter(htmlFile, true);
-									BufferedWriter conexao = new BufferedWriter(fw);
-									conexao.write(str);
-									conexao.newLine();
-									conexao.close();
-
-									List<Element> p = HTMLWorker.parseToList(new FileReader(htmlFile.getPath()),
-											styles);
-
-									for (int k = 0; k < p.size(); ++k) {
-										if (p.get(k) instanceof Paragraph) {
-											Paragraph att = (Paragraph) p.get(k);
-											if (att.getContent().contains("||IMAGE||")) {
-												String img = allMatches.poll();
-												if (img != null) {
-													Image image = Image.getInstance(new URL(
-															img.replaceAll("<img src=\"", "").replaceAll("\">", "")));
-													float scaler = ((document.getPageSize().getWidth()
-															- document.leftMargin() - document.rightMargin())
-															/ image.getWidth()) * 100;
-													image.scalePercent(scaler * 0.4f);
-													image.setAlignment(Element.ALIGN_CENTER);
-													document.add(image);
-												}
-											} else {
-												att.setFirstLineIndent(firstLineIndent);
-												document.add(att);
-
-											}
-										} else if (p.get(k).getClass().getName().equals("com.itextpdf.text.List")) {
-											com.itextpdf.text.List att = (com.itextpdf.text.List) p.get(k);
-											att.setIndentationLeft(firstLineIndent);
-											document.add(att);
-										}
-									}
-									lastAttWasPlan = false;
-									htmlFile.delete();
-								}
-							} else if (a.getType().equals(TableField.class.getCanonicalName())) {
-
-								TableFields tf = fieldsBS.tableFieldsByAttribute(a.getId(), true, false);
-								List<TableStructure> tabStructList = fieldsBS.listTableStructureByFields(tf);
-								List<TableInstance> tabInstList = fieldsBS.listTableInstanceByFields(tf,false);
-								if (!tabInstList.isEmpty()) {
-									if (lastAttWasPlan) {
-										document.setPageSize(PageSize.A4);
-										document.newPage();
-									}
-									String attName = a.getName();
-									if (!subSecTitlePrinted) {
-										subSecIndex++;
-										Chunk c = new Chunk(secIndex + "." + subSecIndex + ". " + subSecName, titulo);
-										c.setGenericTag(secIndex + "." + subSecIndex + ". " + subSecName);
-										Paragraph subSecTitle = new Paragraph(c);
-										subSecTitle.setLeading(interLineSpacing);
-										subSecTitle.setSpacingAfter(paragraphSpacing);
-										subSecTitle.setSpacingBefore(paragraphSpacing);
-										document.add(subSecTitle);
-										subSecTitlePrinted = true;
-									}
-									if (!attName.equals(subSecName)) {
-										Paragraph attTitle = new Paragraph(attName, texto);
-										attTitle.setAlignment(Element.ALIGN_CENTER);
-										attTitle.setLeading(interLineSpacing);
-										attTitle.setSpacingAfter(paragraphSpacing);
-										attTitle.setSpacingBefore(paragraphSpacing);
-										document.add(attTitle);
-									}
-									PdfPTable table = returnPdfPTable(tabStructList, tabInstList, false, false);
-									document.add(table);
-									lastAttWasPlan = false;
-								}
-							} else if (a.getType().equals(ScheduleField.class.getCanonicalName())) {
-								List<ScheduleInstance> schInstList = this.fieldsBS
-										.retrieveScheduleInstanceByAttribute(a.getId(), true);
-
-								if (!schInstList.isEmpty()) {
-									if (lastAttWasPlan) {
-										document.setPageSize(PageSize.A4);
-										document.newPage();
-									}
-									if (!subSecTitlePrinted) {
-										subSecIndex++;
-										Chunk c = new Chunk(secIndex + "." + subSecIndex + ". " + subSecName, titulo);
-										c.setGenericTag(secIndex + "." + subSecIndex + ". " + subSecName);
-										Paragraph subSecTitle = new Paragraph(c);
-										subSecTitle.setLeading(interLineSpacing);
-										subSecTitle.setSpacingAfter(paragraphSpacing);
-										subSecTitle.setSpacingBefore(paragraphSpacing);
-										document.add(subSecTitle);
-										subSecTitlePrinted = true;
-									}
-									String attName = a.getName();
-									if (!attName.equals(subSecName)) {
-										Paragraph attTitle = new Paragraph(attName, texto);
-										attTitle.setAlignment(Element.ALIGN_CENTER);
-										attTitle.setLeading(interLineSpacing);
-										attTitle.setSpacingAfter(paragraphSpacing);
-										attTitle.setSpacingBefore(paragraphSpacing);
-										document.add(attTitle);
-									}
-									PdfPTable table = new PdfPTable(3);
-									table.getDefaultCell();
-									PdfPCell c = new PdfPCell(new Paragraph("Atividade", texto));
-									c.setHorizontalAlignment(Element.ALIGN_CENTER);
-									c.setBackgroundColor(headerBgColor);
-									table.addCell(c);
-									c = new PdfPCell(new Paragraph("Início", texto));
-									c.setHorizontalAlignment(Element.ALIGN_CENTER);
-									c.setBackgroundColor(headerBgColor);
-
-									table.addCell(c);
-									c = new PdfPCell(new Paragraph("Fim", texto));
-									c.setHorizontalAlignment(Element.ALIGN_CENTER);
-									c.setBackgroundColor(headerBgColor);
-									table.addCell(c);
-
-									SimpleDateFormat sdf = new SimpleDateFormat("dd/MM/yyyy");
-									for (ScheduleInstance sch : schInstList) {
-										table.addCell(new Paragraph(sch.getDescription(), texto));
-										table.addCell(new Paragraph(sdf.format(sch.getBegin()), texto));
-										table.addCell(new Paragraph(sdf.format(sch.getEnd()), texto));
-									}
-									document.add(table);
-									lastAttWasPlan = false;
-								}
-							} else if (a.getType().equals(SelectPlan.class.getCanonicalName())) {
-								if (a.getValue() != null) {
-									Plan plan = planBS.retrieveById(Long.parseLong(a.getValue()));
-									// LOGGER.info("2 - " + plan.getName());
-									List<PdfPTable> planTableList = this.generatePDFplanTable(plan);
-									boolean first = true;
-									for (PdfPTable planTable : planTableList) {
-										if (!lastAttWasPlan) {
-											document.setPageSize(PageSize.A4.rotate());
-										}
-										document.newPage();
-										if (first) {
-											if (!subSecTitlePrinted) {
-												subSecIndex++;
-												Chunk c = new Chunk(secIndex + "." + subSecIndex + ". " + subSecName,
-														titulo);
-												c.setGenericTag(secIndex + "." + subSecIndex + ". " + subSecName);
-												Paragraph subSecTitle = new Paragraph(c);
-												document.add(subSecTitle);
-												subSecTitlePrinted = true;
-											}
-											Paragraph attTitle = new Paragraph(plan.getName(), texto);
-											attTitle.setAlignment(Element.ALIGN_CENTER);
-											document.add(attTitle);
-											first = false;
-											lastAttWasPlan = true;
-										}
-										document.add(planTable);
-									}
-									// document.setPageSize(PageSize.A4);
-									// document.newPage();
-								}
-							}
-						}
-					}
-				}
-			}
-		}
-		if (havePreText)
-			preTextDocument.close();
-		if (haveContent)
-			document.close();
-
-		summaryDocument.setPageSize(PageSize.A4);
-		// Margens Superior e esquerda: 3 cm Inferior e direita: 2 cm
-		summaryDocument.setMargins(85.0394f, 56.6929f, 85.0394f, 56.6929f);
-		summaryDocument.open();
-
-		Paragraph summaryTitle = new Paragraph("Sumário", titulo);
-		summaryTitle.setLeading(interLineSpacing);
-		summaryTitle.setSpacingAfter(paragraphSpacing);
-		summaryTitle.setSpacingBefore(paragraphSpacing);
-		summaryDocument.add(summaryTitle);
-
-		Chunk dottedLine = new Chunk(new DottedLineSeparator());
-		List<SimpleEntry<String, SimpleEntry<String, Integer>>> entries = event.getTOC();
-		Paragraph p;
-		int summaryCountPages = 0;
-		for (SimpleEntry<String, SimpleEntry<String, Integer>> entry : entries) {
-			// LOGGER.info(entry.getKey());
-			Chunk chunk = new Chunk(entry.getKey(), titulo);
-			SimpleEntry<String, Integer> value = entry.getValue();
-			chunk.setAction(PdfAction.gotoLocalPage(value.getKey(), false));
-			p = new Paragraph(chunk);
-			p.add(dottedLine);
-			chunk = new Chunk(String.valueOf(value.getValue()), titulo);
-			chunk.setAction(PdfAction.gotoLocalPage(value.getKey(), false));
-			p.add(chunk);
-			summaryDocument.add(p);
-		}
-		summaryDocument.close();
-		PdfReader summaryAux = new PdfReader(summaryPdfFile.getPath());
-		PdfReader cover = new PdfReader(coverPdfFile.getPath());
-		summaryCountPages = summaryAux.getNumberOfPages() + cover.getNumberOfPages();
-		PdfReader preText;
-		if (havePreText) {
-			preText = new PdfReader(preTextPdfFile.getPath());
-			summaryCountPages += preText.getNumberOfPages();
-		}
-
-		com.itextpdf.text.Document finalSummaryDocument = new com.itextpdf.text.Document();
-		PdfWriter.getInstance(finalSummaryDocument, new FileOutputStream(finalSummaryPdfFile));
-		// Formato A4 do documento
-		finalSummaryDocument.setPageSize(PageSize.A4);
-		// Margens Superior e esquerda: 3 cm Inferior e direita: 2 cm
-		finalSummaryDocument.setMargins(85.0394f, 56.6929f, 85.0394f, 56.6929f);
-		finalSummaryDocument.open();
-
-		finalSummaryDocument.add(summaryTitle);
-		for (SimpleEntry<String, SimpleEntry<String, Integer>> entry : entries) {
-			// LOGGER.info(entry.getKey());
-			Chunk chunk = new Chunk(entry.getKey(), titulo);
-			SimpleEntry<String, Integer> value = entry.getValue();
-			chunk.setAction(PdfAction.gotoLocalPage(value.getKey(), false));
-			p = new Paragraph(chunk);
-			p.add(dottedLine);
-			chunk = new Chunk(String.valueOf(value.getValue() + summaryCountPages), titulo);
-			chunk.setAction(PdfAction.gotoLocalPage(value.getKey(), false));
-			p.add(chunk);
-			finalSummaryDocument.add(p);
-		}
-		finalSummaryDocument.close();
-
-		com.itextpdf.text.Document newDocument = new com.itextpdf.text.Document();
-
-		PdfImportedPage page;
-		int n;
-		PdfCopy copy = new PdfCopy(newDocument, new FileOutputStream(destinationFile.getPath()));
-		newDocument.open();
-
-		PdfReader summary = new PdfReader(finalSummaryPdfFile.getPath());
-		PdfReader content;
-		// int unnumberedPgsCount = summaryCountPages;
-		// CAPA
-		n = cover.getNumberOfPages();
-		// unnumberedPgsCount += n;
-		for (int i = 0; i < n;) {
-			page = copy.getImportedPage(cover, ++i);
-			copy.addPage(page);
-		}
-		if (havePreText) {
-			preText = new PdfReader(preTextPdfFile.getPath());
-			// SEÇÕES PRE TEXTUAIS
-			n = preText.getNumberOfPages();
-			// unnumberedPgsCount += n;
-			for (int i = 0; i < n;) {
-				page = copy.getImportedPage(preText, ++i);
-				copy.addPage(page);
-			}
-		}
-		if (haveContent) {
-			// SUMÁRIO
-			n = summary.getNumberOfPages();
-			for (int i = 0; i < n;) {
-				page = copy.getImportedPage(summary, ++i);
-				copy.addPage(page);
-			}
-			content = new PdfReader(contentFile.getPath());
-			// CONTEÚDO
-			n = content.getNumberOfPages();
-			for (int i = 0; i < n;) {
-				page = copy.getImportedPage(content, ++i);
-				copy.addPage(page);
-			}
-		}
-
-		newDocument.close();
-
-		manipulatePdf(destinationFile.getPath(), finalPdfFile.getPath(), newDocument, summaryCountPages);
-		InputStream inpStr = new FileInputStream(finalPdfFile);
-		return inpStr;
 	}
 
 	/**
@@ -2948,7 +2069,6 @@ public class DocumentBS extends HibernateBusiness {
 		reader.close();
 	}
 	
-	
 	/**
 	 * Lista com os valores dos campos
 	 * 
@@ -2989,6 +2109,15 @@ public class DocumentBS extends HibernateBusiness {
 				.addOrder(Order.asc("id"));
 		Schedule schedule = (Schedule) criteria.uniqueResult();
 		return schedule;
+	}
+
+	public List<AggregateIndicator> listAggIndbyLevelInstance(StructureLevelInstance levelInstance) {
+		Criteria criteria = this.dao.newCriteria(AggregateIndicator.class);
+		criteria.add(Restrictions.eq("deleted", false));
+		criteria.add(Restrictions.eq("indicator", levelInstance));
+		List<AggregateIndicator> levelList = this.dao.findByCriteria(criteria, AggregateIndicator.class);
+		
+		return levelList;
 	}
 
 }
