@@ -1,20 +1,20 @@
 
 import React from "react";
+import { ToastContainer } from 'react-toastr';
 
-import MainMenu from "forpdi/jsx/MainMenu.jsx";
-import TopBar from "forpdi/jsx/TopBar.jsx";
+import ForPDITopBar from "forpdi/jsx/TopBar.jsx";
+import ForRiscoTopBar from "forpdi/jsx_forrisco/TopBar.jsx";
+import PlataformaForTopbar from "forpdi/jsx/PlataformaForTopbar.jsx";;
 
 import AccessLevels from "forpdi/jsx/core/store/AccessLevels.json";
 import UserSession from "forpdi/jsx/core/store/UserSession.jsx";
-import Login from "forpdi/jsx/core/view/user/Login.jsx";
 
 import LoadingGauge from "forpdi/jsx/core/widget/LoadingGauge.jsx";
 
-var ReactToastr = require("react-toastr");
-var {ToastContainer} = ReactToastr; 
-var ToastMessageFactory = React.createFactory(ReactToastr.ToastMessage);
-
 export default React.createClass({
+	contextTypes: {
+		router: React.PropTypes.object.isRequired,
+	},
 	childContextTypes: {
 		accessLevel: React.PropTypes.number,
 		accessLevels: React.PropTypes.object,
@@ -38,7 +38,7 @@ export default React.createClass({
 		};
 	},
 	addAlertError(msg) {
-		this.refs.container.clear(); 
+		this.refs.container.clear();
 		this.refs.container.error(
 			msg,null, {
 				timeOut: 5000,
@@ -69,27 +69,34 @@ export default React.createClass({
 		};
 	},
 	componentDidMount() {
-		var me = this;
-		//$("[data-toggle=tooltip]").tooltip();
+		if (!this.state.loading && this.state.accessLevel === 0)
+			this.context.router.push("/login");
 		UserSession.on("login", () => {
-			me.setState({
+			this.setState({
 				accessLevel: UserSession.get("accessLevel") || 0,
 				permissions: UserSession.get("permissions") || []
 			});
-		}, me);
+		}, this);
 		UserSession.on("logout", () => {
-			me.setState({
+			this.setState({
 				accessLevel: 0,
 				permissions: []
 			});
-		}, me);
-		UserSession.on("loaded", () => {me.setState({loading: false})}, me);
+		}, this);
+		UserSession.on("loaded", () => {
+			this.setState({ loading: false });
+		}, this);
 	},
 	componentWillUnmount() {
 		UserSession.off(null, null, this);
 	},
-	componentDidUpdate() {
-		//$("[data-toggle=tooltip]").tooltip();
+	getTopbar() {
+		if (this.props.location.pathname.includes("forrisco")) {
+			return <ForRiscoTopBar />;
+		} else if (this.props.location.pathname.includes("app-select")) {
+			return <PlataformaForTopbar />;
+		}
+		return <ForPDITopBar />;
 	},
 	render() {
 		if (this.state.loading) {
@@ -97,15 +104,10 @@ export default React.createClass({
 		}
 		return (
 			<main className='fpdi-app-container'>
-				<TopBar />
-				<div className="fpdi-app-body">
-					<MainMenu {...this.props} />
-					<ToastContainer ref="container"					
-						className="toast-top-center" />
-					<div className="fpdi-app-content">
-						  {this.state.accessLevel == 0 ? <Login /> : this.props.children}
-					</div>
-				</div>
+				{ this.getTopbar() }
+				<ToastContainer ref="container"
+					className="toast-top-center" />
+				{this.props.children}
 			</main>
 		);
 	}
