@@ -1,14 +1,12 @@
 import React from "react";
-import {Link} from "react-router";
+import { Link } from "react-router";
 import Form from "forpdi/jsx/core/widget/form/Form.jsx";
 import UserSession from "forpdi/jsx/core/store/UserSession.jsx";
 import LoadingGauge from "forpdi/jsx/core/widget/LoadingGauge.jsx";
-import Modal from "forpdi/jsx/core/widget/Modal.jsx";
-//import Toastr from 'toastr';
 
 import Messages from "forpdi/jsx/core/util/Messages.jsx";
 
-import AppLogo from "forpdi/img/logoLogin.png";
+import AppLogo from "forpdi/img/plataforma-for-logo.svg";
 import Validation from 'forpdi/jsx/core/util/Validation.jsx';
 
 var Validate = Validation.validate;
@@ -19,6 +17,7 @@ export default React.createClass({
 	contextTypes: {
 		toastr: React.PropTypes.object.isRequired
 	},
+
 	getInitialState() {
 		return {
 			loaded: !UserSession.get("loading"),
@@ -37,6 +36,7 @@ export default React.createClass({
 			}]
 		};
 	},
+
 	onSubmit(data) {
 		Validate.validationLogin(this.refs["login"]);
 
@@ -45,50 +45,67 @@ export default React.createClass({
 			data: data
 		});
 	},
+
 	componentWillMount() {
 		var me = this;
 		UserSession.on("login", model => {
 			var url = window.location.href.split("#");
 			var path = url[1].split("?");
 			if (path[0] == "/login")
-				location.assign("#/home");
+				location.assign("#/app-select");
 		}, me);
 		UserSession.on("loaded", () => {
-			me.setState({loaded: true});
+			me.setState({ loaded: true });
 		}, me);
 
 	},
+
 	componentDidMount() {
 		if (!!UserSession.get("logged")) {
-			location.assign("#/home");
+			location.assign("#/forrisco/home");
 		} else {
-			this.setState({loaded: true});
+			this.setState({ loaded: true });
 		}
 	},
+
 	componentWillUnmount() {
 		UserSession.off(null, null, this);
 	},
+
 	render() {
 		if (!this.state.loaded) {
 			return <LoadingGauge />;
 		}
+
 		return (
 			<div className="container-fluid">
 				<div className="row">
 					<div className="col-xs-12 text-center">
 						<div className="fpdi-login-header">
-							<img className="fpdi-login-brand" src={AppLogo} alt={Messages.getEditable("label.forPdiLogo","fpdi-nav-label")} />
-							<h3 className="fpdi-login-subtitle">{Messages.get("label.login.titleComplement")}<br/>{Messages.getEditable("label.login.title","fpdi-nav-label")}</h3>
+							<img
+								className="fpdi-login-brand"
+								src={AppLogo}
+								alt={Messages.getEditable("label.forRiscoLogo","fpdi-nav-label")}
+							/>
+							<center>
+								<h3 className="frisco-login-subtitle">
+									{Messages.get("label.login.titlePlatformComplement")}<br/>
+									{/* Messages.getEditable("label.login.title","fpdi-nav-label") */}
+								</h3>
+							</center>
 						</div>
 					</div>
 				</div>
 
-
 		    <div className="row">
-				<div className="col-md-4 col-md-offset-4">	
-					<div className="fpdi-card-login">		
-						<div className="panel panel-default">
-						  <div className="panel-heading"><p className="fpdi-login-title">{Messages.getEditable("label.login","fpdi-nav-label")}</p></div>
+					<div className="col-md-4 col-md-offset-4">
+						<div className="fpdi-card-login">
+							<div className="panel panel-default">
+							  <div className="panel-heading">
+									<p className="fpdi-login-title">
+										{Messages.getEditable("label.login","fpdi-nav-label")}
+									</p>
+								</div>
 							  <div className="panel-body">
 									<div className="fpdi-login-body">
 										<VerticalForm
@@ -103,7 +120,7 @@ export default React.createClass({
 											id="login-form"
 										/>
 									</div>
-									
+
 									<div className="fpdi-login-footer">
 										<div className="row">
 											<div className="col-md-12 text-center marginBottom10">
@@ -112,18 +129,15 @@ export default React.createClass({
 										</div>
 									</div>
 							  </div>
+							</div>
+						</div>
+
+						<div className="fpdi-browsers-info">
+							{Messages.getEditable("label.infoBrowsers","fpdi-nav-label")}<br/>
+							<i>{Messages.getEditable("label.browsers","fpdi-nav-label")}</i>
 						</div>
 					</div>
-
-					<div className="fpdi-browsers-info">
-						{Messages.getEditable("label.infoBrowsers","fpdi-nav-label")}<br/>
-						<i>{Messages.getEditable("label.browsers","fpdi-nav-label")}</i>
-					</div>
 				</div>
-			</div>
-
-
-
 			</div>
 		);
 	}
