@@ -1,9 +1,9 @@
-
 import React from "react";
-import {Link} from 'react-router';
-import UserStore from "forpdi/jsx/core/store/User.jsx";
+import { Link } from 'react-router';
 import string from 'string';
-import EditableDataView from "forpdi/jsx/core/widget/form/EditableDataView.jsx";
+import UserStore from "forpdi/jsx/core/store/User.jsx";
+import _ from "underscore";
+
 import Form from "forpdi/jsx/core/widget/form/Form.jsx";
 import LoadingGauge from "forpdi/jsx/core/widget/LoadingGauge.jsx";
 import Modal from "forpdi/jsx/core/widget/Modal.jsx";
@@ -11,13 +11,10 @@ import AccessLevels from "forpdi/jsx/core/store/AccessLevels.json";
 import Logo from 'forpdi/img/foto_padrao.png';
 import FileStore from "forpdi/jsx/core/store/File.jsx"
 import UserSession from "forpdi/jsx/core/store/UserSession.jsx";
-import _ from "underscore";
 import Validation from 'forpdi/jsx/core/util/Validation.jsx';
 import NotificationMessageUser from "forpdi/jsx/core/view/user/NotificationMessageUser.jsx";
 import Messages from "forpdi/jsx/core/util/Messages.jsx";
-
-
-//import Toastr from 'toastr';
+import UserPermissions from "forpdi/jsx/core/widget/user/UserPermissions.jsx";
 
 var Validate = Validation.validate;
 
@@ -29,11 +26,10 @@ export default React.createClass({
 		router: React.PropTypes.object,
 		roles: React.PropTypes.object.isRequired,
 		toastr: React.PropTypes.object.isRequired,
-		permissions: React.PropTypes.array.isRequired
 	},
 	getInitialState() {
 		return {
-			dataUser:null,
+			dataUser: null,
 			model: null,
 			isEditUser: false,
 			fields: this.getFields(),
@@ -41,17 +37,13 @@ export default React.createClass({
 			infoDocument: false,
 			infoGoal: false,
 			user: UserSession.get("user"),
-			permissions: [],
-			detailed: [],
-			selectedPermissions: [],
-			editingPermission: false,
 			passwordChange: false,
 			loading: !!this.props.params.modelId,
 			modelId: this.props.params.modelId,
 			fieldsMessage: this.getFieldsMessage(),
 			page: 1,
 			totalMessages: 0,
-			hideShowMoreMessages:false
+			hideShowMoreMessages: false
 			//url: this.props.location.pathname,
 		};
 	},
@@ -65,26 +57,15 @@ export default React.createClass({
 		// 	return;
 		// }
 
-		UserStore.on("retrieve-permissions", (model) => {
-			model.data.map((item, idx) => {
-				if(item.granted)
-					this.state.selectedPermissions.push(idx);
-			});
-			me.setState({
-				permissions: model.data
-			});
-		}, me);
-
 		UserStore.on("retrieve", (model) => {
 			me.setState({
 				model: model.attributes,
 			});
-			this.getPermissionsList();
 
 			UserSession.dispatch({
-			action: UserSession.ACTION_LIST_MESSAGES,
+				action: UserSession.ACTION_LIST_MESSAGES,
 				data: {
-					userId:this.state.model.id,
+					userId: this.state.model.id,
 					limit: 3
 				}
 			});
@@ -108,9 +89,7 @@ export default React.createClass({
 			data: this.props.params.modelId
 		});
 
-		UserSession.on("retrieve-showMoreMessages",(model) => {
-			//console.log(model);
-
+		UserSession.on("retrieve-showMoreMessages", (model) => {
 			if ((this.state.totalMessages + model.data.length) >= model.total) {
 				me.setState({
 					hideShowMoreMessages: true
@@ -119,10 +98,10 @@ export default React.createClass({
 			me.setState({
 				totalMessages: model.data.length + this.state.totalMessages
 			});
-		},me);
+		}, me);
 
 
-		UserSession.on("retrieve-messages",(model) => {
+		UserSession.on("retrieve-messages", (model) => {
 			if (model.data.length >= model.total || model.data.length <= 0) {
 				me.setState({
 					hideShowMoreMessages: true
@@ -131,7 +110,7 @@ export default React.createClass({
 			me.setState({
 				totalMessages: model.data.length
 			});
-		},me);
+		}, me);
 
 
 		UserStore.on("editUser", (store) => {
@@ -150,36 +129,19 @@ export default React.createClass({
 			}
 		}, me);
 
-		UserSession.on("sendMessage",(store) => {
+		UserSession.on("sendMessage", (store) => {
 			this.context.toastr.addAlertSuccess(Messages.get("label.success.sendMessage"));
 			this.onClear();
 
 			UserSession.dispatch({
-			action: UserSession.ACTION_LIST_MESSAGES,
+				action: UserSession.ACTION_LIST_MESSAGES,
 				data: {
-					userId:this.state.model.id,
+					userId: this.state.model.id,
 					limit: 3
 				}
 			});
 
-		},me);
-
-		UserStore.on("permissions-saved", (store) => {
-			me.setState({
-				editingPermission: false
-			});
-			this.context.toastr.addAlertSuccess(Messages.get("label.success.editedSuccessfully"));
-			me.updateLoadingState();
 		}, me);
-	},
-
-	getPermissionsList(){
-		UserStore.dispatch({
-			action: UserStore.ACTION_LIST_PERMISSIONS,
-			data: {
-				userId: this.state.model.id
-			}
-		})
 	},
 
 	componentWillUnmount() {
@@ -191,9 +153,9 @@ export default React.createClass({
 			name: "name",
 			type: "text",
 			placeholder: "",
-			label: Messages.getEditable("label.name","fpdi-nav-label"),
-			maxLength:255,
-			value: model  ? model.name:null,
+			label: Messages.getEditable("label.name", "fpdi-nav-label"),
+			maxLength: 255,
+			value: model ? model.name : null,
 			required: true,
 			onKeyPress: this.onlyLetter,
 			onPaste: this.onlyLetterPaste
@@ -202,9 +164,9 @@ export default React.createClass({
 			name: "email",
 			type: "email",
 			placeholder: "",
-			label: Messages.getEditable("label.email","fpdi-nav-label"),
-			maxLength:255,
-			value:  model ? model.email:null,
+			label: Messages.getEditable("label.email", "fpdi-nav-label"),
+			maxLength: 255,
+			value: model ? model.email : null,
 			required: true,
 			disabled: this.context.roles.SYSADMIN || (model && UserSession.get("user").id == model.id) ? false : true,
 			disableMsg: Messages.get("label.disabled.cpf")
@@ -214,8 +176,8 @@ export default React.createClass({
 			name: "cpf",
 			type: "cpf",
 			placeholder: "",
-			label: Messages.getEditable("label.cpf","fpdi-nav-label"),
-			value:  model ? model.cpf:null,
+			label: Messages.getEditable("label.cpf", "fpdi-nav-label"),
+			value: model ? model.cpf : null,
 			required: true,
 			disabled: this.context.roles.SYSADMIN || (model && UserSession.get("user").id == model.id) ? false : true,
 			disableMsg: Messages.get("label.disabled.cpf")
@@ -226,9 +188,9 @@ export default React.createClass({
 			type: "date",
 			required: true,
 			placeholder: "",
-			label: Messages.getEditable("label.birthdate","fpdi-nav-label"),
+			label: Messages.getEditable("label.birthdate", "fpdi-nav-label"),
 			onChange: this.onStartDateChange,
-			value: 	model ? model.birthdate:null,
+			value: model ? model.birthdate : null,
 			required: true
 		},
 
@@ -236,25 +198,25 @@ export default React.createClass({
 			name: "cellphone",
 			type: "tel",
 			placeholder: "",
-			label: Messages.getEditable("label.cellphone","fpdi-nav-label"),
-			value:  model ? model.cellphone:null,
+			label: Messages.getEditable("label.cellphone", "fpdi-nav-label"),
+			value: model ? model.cellphone : null,
 			required: true
 		},
 		{
 			name: "phone",
 			type: "tel",
 			placeholder: "",
-			label: Messages.getEditable("label.phone","fpdi-nav-label"),
-			value: 	model? model.phone:null,
+			label: Messages.getEditable("label.phone", "fpdi-nav-label"),
+			value: model ? model.phone : null,
 		},
 
 		{
 			name: "department",
 			type: "text",
 			placeholder: "",
-			label: Messages.getEditable("label.department","fpdi-nav-label"),
-			maxLength:255,
-			value:  model ? model.department:null,
+			label: Messages.getEditable("label.department", "fpdi-nav-label"),
+			maxLength: 255,
+			value: model ? model.department : null,
 		},
 		{
 			name: 'accessLevel',
@@ -263,14 +225,14 @@ export default React.createClass({
 			placeholder: Messages.get("label.userLevel"),
 			label: Messages.get("label.userType"),
 			required: true,
-			value: model ? model.accessLevel :null,
+			value: model ? model.accessLevel : null,
 			displayField: 'name',
 			valueField: 'accessLevel',
 			options: this.context.roles.SYSADMIN ?
 				AccessLevels.list.filter((level) => {
 					return level.accessLevel <= this.context.accessLevel
 				})
-			:
+				:
 				AccessLevels.listNoSysAdm.filter((level) => {
 					return level.accessLevel <= this.context.accessLevel;
 				})
@@ -279,21 +241,20 @@ export default React.createClass({
 			name: "newPassword",
 			type: "password",
 			placeholder: "",
-			label: Messages.getEditable("label.newPassword","fpdi-nav-label"),
+			label: Messages.getEditable("label.newPassword", "fpdi-nav-label"),
 			value: "",
 			required: false
 		}]
 
 	},
 
-
-	getFieldsMessage (model) {
+	getFieldsMessage(model) {
 		return [{
 			name: "assunto",
 			type: "subject",
 			placeholder: "",
-			label: Messages.getEditable("label.subject","fpdi-nav-label"),
-			maxLength:70,
+			label: Messages.getEditable("label.subject", "fpdi-nav-label"),
+			maxLength: 70,
 			value: null,
 			required: true
 		},
@@ -301,8 +262,8 @@ export default React.createClass({
 			name: "mensagem",
 			type: "message",
 			placeholder: "",
-			label: Messages.getEditable("label.msg","fpdi-nav-label"),
-			maxLength:255,
+			label: Messages.getEditable("label.msg", "fpdi-nav-label"),
+			maxLength: 255,
 			value: null,
 			required: true
 		}]
@@ -323,15 +284,15 @@ export default React.createClass({
 	},
 
 	showMoreMessages() {
-		var newPage = this.state.page+1;
+		var newPage = this.state.page + 1;
 		UserSession.dispatch({
-        	action: UserSession.ACTION_LIST_MESSAGES,
-        	data: {
-        		userId:this.state.model.id,
-    			limit: 3,
-    			page: newPage
-    		}
-    	});
+			action: UserSession.ACTION_LIST_MESSAGES,
+			data: {
+				userId: this.state.model.id,
+				limit: 3,
+				page: newPage
+			}
+		});
 		this.setState({
 			page: newPage
 		});
@@ -347,20 +308,20 @@ export default React.createClass({
 		} else {
 			UserSession.dispatch({
 				action: UserSession.ACTION_SEND_MESSAGE,
-					data: {
-						subject:data.assunto,
-						message:data.mensagem,
-						userId:this.state.model.id
-					}
-				});
+				data: {
+					subject: data.assunto,
+					message: data.mensagem,
+					userId: this.state.model.id
+				}
+			});
 		}
 
 	},
 
-	onStartDateChange(data){
+	onStartDateChange(data) {
 		var model = this.state.model;
 		this.setState({
-			model:model,
+			model: model,
 		});
 		this.refs.editUser.refs.birthdate.props.fieldDef.value = data.format('DD/MM/YYYY');
 	},
@@ -372,21 +333,21 @@ export default React.createClass({
 		});
 	},
 
-	validarCPF(cpf){
-		if(cpf.length != 11 || cpf.replace(eval('/'+cpf.charAt(1)+'/g'),'') == '') {
-		    return false;
+	validarCPF(cpf) {
+		if (cpf.length != 11 || cpf.replace(eval('/' + cpf.charAt(1) + '/g'), '') == '') {
+			return false;
 		} else {
-	   	    var d;
-	   	    var c;
-	        for(var n=9; n<11; n++) {
-	            for(d=0, c=0; c<n; c++)
-	            	d += cpf.charAt(c) * ((n + 1) - c);
-	            d = ((10 * d) % 11) % 10;
-	            if(cpf.charAt(c) != d)
-	            	return false;
-	        }
-	      return true;
-	   }
+			var d;
+			var c;
+			for (var n = 9; n < 11; n++) {
+				for (d = 0, c = 0; c < n; c++)
+					d += cpf.charAt(c) * ((n + 1) - c);
+				d = ((10 * d) % 11) % 10;
+				if (cpf.charAt(c) != d)
+					return false;
+			}
+			return true;
+		}
 	},
 
 
@@ -410,11 +371,11 @@ export default React.createClass({
 					UserStore.dispatch({
 						action: UserStore.ACTION_UPDATE_USER,
 						data: {
-							user:data,
+							user: data,
 							newPassword: data.newPassword == "" ? null : data.newPassword
 						}
 					});
-				},msg,this.refreshCancel);
+				}, msg, this.refreshCancel);
 			} else {
 				var msg = Messages.get("label.msgCreateUser");
 				Modal.confirmCustom(() => {
@@ -423,7 +384,7 @@ export default React.createClass({
 						action: UserStore.ACTION_CREATE,
 						data: data
 					});
-				},msg,this.refreshCancel);
+				}, msg, this.refreshCancel);
 
 
 				/*UserStore.dispatch({
@@ -435,27 +396,25 @@ export default React.createClass({
 
 	},
 
-
-
-	showPlanGoals () {
+	showPlanGoals() {
 		this.setState({
 			infoPlanGoals: !this.state.infoPlanGoals
 		});
 	},
 
-	showDocumen () {
+	showDocumen() {
 		this.setState({
 			infoDocument: !this.state.infoDocument
 		});
 	},
 
-	showGoal () {
+	showGoal() {
 		this.setState({
 			infoGoal: !this.state.infoGoal
 		});
 	},
 
-	refreshCancel () {
+	refreshCancel() {
 		Modal.hide();
 		//this.setState({
 		//	editUser: false
@@ -463,15 +422,14 @@ export default React.createClass({
 		//});
 	},
 
-
-	refreshAccept () {
+	refreshAccept() {
 		Modal.hide();
 		this.setState({
 			isEditUser: false
 		});
 	},
 
-	onCancel () {
+	onCancel() {
 		var deleteFieldsPassword;
 		var tamDeleteFieldsPassword;
 
@@ -479,25 +437,25 @@ export default React.createClass({
 		tamDeleteFieldsPassword = deleteFieldsPassword.length - 1;
 
 		if (deleteFieldsPassword[tamDeleteFieldsPassword].name == "newPasswordTwo") {
-				deleteFieldsPassword.pop();
-				deleteFieldsPassword.pop();
-				deleteFieldsPassword.pop();
-				deleteFieldsPassword.push({
-					name: "senha",
-					type: "changePassword",
-					placeholder: Messages.get("label.changePassword"),
-					onClick: this.changePassword
-				})
+			deleteFieldsPassword.pop();
+			deleteFieldsPassword.pop();
+			deleteFieldsPassword.pop();
+			deleteFieldsPassword.push({
+				name: "senha",
+				type: "changePassword",
+				placeholder: Messages.get("label.changePassword"),
+				onClick: this.changePassword
+			})
 		}
 
 		this.setState({
-			fields:deleteFieldsPassword
+			fields: deleteFieldsPassword
 		})
 		var msg = Messages.get("label.msgEdit");
-		Modal.confirmCancelCustom(this.refreshAccept,msg,this.refreshCancel);
+		Modal.confirmCancelCustom(this.refreshAccept, msg, this.refreshCancel);
 	},
 
-	confirmEdit () {
+	confirmEdit() {
 		this.setState({
 			isEditUser: true
 		});
@@ -513,11 +471,11 @@ export default React.createClass({
 					{Messages.get("label.selectDocument")}
 				</p>
 			</div>),
-			FileStore.url+"/upload",
+			FileStore.url + "/upload",
 			"image/*",
 			formatsBlocked,
 			(resp) => {
-				me.setState({url: resp.message});
+				me.setState({ url: resp.message });
 				Modal.hide();
 				UserStore.dispatch({
 					action: UserStore.ACTION_UPDATE_PICTURE_EDIT_USER,
@@ -530,20 +488,20 @@ export default React.createClass({
 			},
 			(resp) => {
 				Modal.hide();
-				me.setState({error: resp.message});
+				me.setState({ error: resp.message });
 			},
 			"jpg, jpeg, gif, png, svg."
 		);
 	},
 
-	showDescription(description){
+	showDescription(description) {
 		var items = description.split(",");
-		return(
+		return (
 			<div className="user-permission-detail">
 				<ul>
 					{items.map((item, idx) => {
-						return(
-							<li key={"item-"+idx}>{item}</li>
+						return (
+							<li key={"item-" + idx}>{item}</li>
 						)
 					})}
 				</ul>
@@ -551,274 +509,182 @@ export default React.createClass({
 		);
 	},
 
-	details(idx){
-		var array = this.state.detailed || [];
-		if(_.contains(array, idx)){
-			array.splice(array.indexOf(idx), 1);
-		} else {
-			array.push(idx);
-		}
-		this.setState({
-			detailed: array
-		});
-	},
-
-	editPermissions(){
-		this.setState({
-			editingPermission: !this.state.editingPermission
-		});
-	},
-
-	selectPermission(idx){
-		var array = this.state.selectedPermissions || [];
-		if(_.contains(array, idx)){
-			this.state.permissions[idx].granted = false;
-			array.splice(array.indexOf(idx), 1);
-		} else {
-			this.state.permissions[idx].granted = true;
-			array.push(idx);
-		}
-		this.setState({
-			selectedPermissions: array
-		});
-	},
-
-	savePermissions(){
-		var list = [];
-		this.state.permissions.map((permission) => {
-			list.push({
-				user: {
-					id: this.state.model.id
-				},
-				permission: permission.type,
-				revoked: !permission.granted
-			});
-		});
-
-		UserStore.dispatch({
-			action: UserStore.ACTION_SAVE_PERMISSIONS,
-			data: {
-				list: list,
-				total: list.length
-			}
-		})
-	},
-
-	onlyLetter(evt){
+	onlyLetter(evt) {
 		var regex = new RegExp("^[a-zA-ZáàâãéèêíïóôõöúçñÁÀÂÃÉÈÍÏÓÔÕÖÚÇÑ ]+$");
-	    var key = String.fromCharCode(!evt.charCode ? evt.which : evt.charCode);
-	    if (!regex.test(key)) {
-	       evt.preventDefault();
-	       return false;
-	    }
+		var key = String.fromCharCode(!evt.charCode ? evt.which : evt.charCode);
+		if (!regex.test(key)) {
+			evt.preventDefault();
+			return false;
+		}
 	},
 
-   onlyLetterPaste(evt){
-   		var regex = new RegExp("^[a-zA-ZáàâãéèêíïóôõöúçñÁÀÂÃÉÈÍÏÓÔÕÖÚÇÑ ]+$");
-	   	var value = evt.clipboardData.getData('Text');
-	   	if (!regex.test(value)) {
-		   evt.preventDefault();
-		   return;
-	   	}
-   },
+	onlyLetterPaste(evt) {
+		var regex = new RegExp("^[a-zA-ZáàâãéèêíïóôõöúçñÁÀÂÃÉÈÍÏÓÔÕÖÚÇÑ ]+$");
+		var value = evt.clipboardData.getData('Text');
+		if (!regex.test(value)) {
+			evt.preventDefault();
+			return;
+		}
+	},
 
 	render() {
-
 		if (this.state.loading) {
 			return <LoadingGauge />;
 		}
 		var birth;
-		//console.log(this.props.location.pathname);
 		if (this.state.model.birthdate) {
 			try {
-				birth=this.state.model.birthdate.split(" ")[0];
+				birth = this.state.model.birthdate.split(" ")[0];
 			}
-			catch(err) {
-				birth=this.state.model.birthdate.getDate() + "/" + this.state.model.birthdate.getMonth+1
-				+ "/" + this.state.model.birthdate.getFullYear();
+			catch (err) {
+				birth = this.state.model.birthdate.getDate() + "/" + this.state.model.birthdate.getMonth + 1
+					+ "/" + this.state.model.birthdate.getFullYear();
 			}
 		}
 
 		return (
-  				<div className="fpdi-profile-user padding40">
-				 	<h1  id = "title-profile-user"> {Messages.getEditable("label.editUser","fpdi-nav-label")} </h1>
-						<div className="marginBottom20">
-							<span>
-								<Link className="fpdi-breadcrumb fpdi-breadcrumbDivisor"
-									to={"/users"}
-									title={"Usuários"}>{Messages.getEditable("label.users","fpdi-nav-label")}</Link>
-								<span className="mdi mdi-chevron-right fpdi-breadcrumbDivisor"></span>
-							</span>
+			<div className="fpdi-profile-user padding40">
+				<h1 id="title-profile-user"> {Messages.getEditable("label.editUser", "fpdi-nav-label")} </h1>
+				<div className="marginBottom20">
+					<span>
+						<Link className="fpdi-breadcrumb fpdi-breadcrumbDivisor"
+							to={"/users"}
+							title={"Usuários"}>{Messages.getEditable("label.users", "fpdi-nav-label")}</Link>
+						<span className="mdi mdi-chevron-right fpdi-breadcrumbDivisor"></span>
+					</span>
 
-							<span className="fpdi-breadcrumb fpdi-selectedOnBreadcrumb">
-								{this.state.model.name}
-							</span>
+					<span className="fpdi-breadcrumb fpdi-selectedOnBreadcrumb">
+						{this.state.model.name}
+					</span>
+				</div>
+				<div className="row">
+
+					<div className="col-sm-3">
+						{this.state.isEditUser == false ?
+
+							<div className="panel panel-default panel-default-user">
+								<div className="panel-heading">{Messages.getEditable("label.userData", "fpdi-nav-label")}
+									{(this.context.roles.ADMIN || _.contains(this.context.permissions,
+										"org.forpdi.core.user.authz.permission.ManageUsersPermission")) ?
+										<span className="mdi mdi-pencil cursorPointer floatRight" onClick={this.confirmEdit} title={Messages.get("label.title.editUserProfile")} />
+										: ""
+									}
+								</div>
+								<div className="panel-body">
+
+									<div className="fpdi-container-profile">
+										<img className="fpdi-image-profile" src={this.state.model.picture == null ? (Logo) : this.state.model.picture} />
+									</div>
+									{(this.context.roles.ADMIN || _.contains(this.context.permissions,
+										"org.forpdi.core.user.authz.permission.ManageUsersPermission") ?
+										<div className="fpdi-container-profile-icon">
+											<span className="mdi mdi-camera mdi-camera-attributes cursorPointer" onClick={this.uploadFile} title={Messages.get("label.title.changeProfileImage")} />
+										</div> : "")}
+
+									<div>
+										<form>
+											<div className="form-group form-profile">
+												<label className="fpdi-text-label"> {Messages.getEditable("label.name", "fpdi-nav-label")} </label>
+												<p id="p-profileUser"> {(this.state.model.name.length > 25) ? (string(this.state.model.name).trim().substr(0, 25).concat("...").toString()) : (this.state.model.name)} </p>
+
+												<label className="fpdi-text-label"> {Messages.getEditable("label.email", "fpdi-nav-label")} </label>
+												<p id="p-profileUser"> {this.state.model.email} </p>
+
+												<label className="fpdi-text-label"> {Messages.getEditable("label.cpf", "fpdi-nav-label")} </label>
+												<p id="p-profileUser"> {this.state.model.cpf} </p>
+
+												<label className="fpdi-text-label"> {Messages.getEditable("label.birthdate", "fpdi-nav-label")} </label>
+												<p id="p-profileUser"> {this.state.model.birthdate == null ? this.state.model.birthdate : this.state.model.birthdate.split(" ")[0]} </p>
+
+												<label className="fpdi-text-label">{Messages.getEditable("label.cellphone", "fpdi-nav-label")}</label>
+												<p id="p-profileUser"> {this.state.model.cellphone} </p>
+
+												<label className="fpdi-text-label">{Messages.getEditable("label.phone", "fpdi-nav-label")}</label>
+												<p id="p-profileUser"> {this.state.model.phone}  </p>
+
+												<label className="fpdi-text-label"> {Messages.getEditable("label.department", "fpdi-nav-label")}</label>
+												<p id="p-profileUser"> {this.state.model.department == null ? this.state.model.department : (this.state.model.department.length > 25) ? (string(this.state.model.department).trim().substr(0, 25).concat("...").toString()) : (this.state.model.department)} </p>
+
+												<label className="fpdi-text-label"> {Messages.getEditable("label.userType", "fpdi-nav-label")} </label>
+												<p id="p-profileUser"> {AccessLevels.mapped[this.state.model.accessLevel]} </p>
+											</div>
+										</form>
+									</div>
+
+								</div>
+							</div>
+
+							:
+
+							<div className="panel panel-default panel-default-user">
+								<div className="panel-heading">{Messages.getEditable("label.userData", "fpdi-nav-label")}</div>
+								<div className="panel-body">
+
+									<div className="fpdi-container-profile">
+										<img className="fpdi-image-profile" src={this.state.model.picture == null ? (Logo) : this.state.model.picture} />
+									</div>
+									{(this.context.roles.ADMIN || _.contains(this.context.permissions,
+										"org.forpdi.core.user.authz.permission.ManageUsersPermission") ?
+										<div className="fpdi-container-profile-icon">
+											<span className="mdi mdi-camera mdi-camera-attributes cursorPointer" onClick={this.uploadFile} title={Messages.get("label.title.changeProfileImage")} />
+										</div> : "")}
+
+									<div className="panel-body">
+										<VerticalForm
+											ref="editUser"
+											onSubmit={this.onSubmit}
+											fields={this.state.fields}
+											store={UserStore}
+											submitLabel={Messages.get("label.submitLabel")}
+											onCancel={this.onCancel}
+										/>
+
+									</div>
+
+								</div>
+
+
+							</div>}
+
+					</div>
+
+					<UserPermissions user={this.state.model} />
+
+					<div className="col-sm-5">
+						<div className="panel panel-default panel-default-user">
+							<div className="panel-heading">{Messages.getEditable("label.sendMessages", "fpdi-nav-label")}</div>
+							<div className="padding5">
+								<div className="panel-body">
+									<VerticalForm
+										ref="messageUser"
+										onSubmit={this.onSubmitMenssage}
+										fields={this.state.fieldsMessage}
+										store={UserStore}
+										submitLabel={Messages.get("label.send")}
+										cancelLabel={Messages.get("label.clean")}
+										onCancel={this.onClear}
+									/>
+								</div>
+							</div>
 						</div>
-						<div className="row">
 
-							<div className="col-sm-3">
-								{this.state.isEditUser == false ?
-
-   	 								<div className="panel panel-default panel-default-user">
-      									<div className="panel-heading">{Messages.getEditable("label.userData","fpdi-nav-label")}
-      										{(this.context.roles.ADMIN  || _.contains(this.context.permissions,
-														"org.forpdi.core.user.authz.permission.ManageUsersPermission")) ?
-      											<span className="mdi mdi-pencil cursorPointer floatRight" onClick={this.confirmEdit} title={Messages.get("label.title.editUserProfile")}/>
-      											: ""
-      										}
-      									</div>
-      										<div className="panel-body">
-
-												<div className="fpdi-container-profile">
-													<img className="fpdi-image-profile" src={this.state.model.picture == null ?(Logo):this.state.model.picture}/>
-												</div>
-												{(this.context.roles.ADMIN  || _.contains(this.context.permissions,
-														"org.forpdi.core.user.authz.permission.ManageUsersPermission") ?
-												<div className="fpdi-container-profile-icon">
-  													<span className="mdi mdi-camera mdi-camera-attributes cursorPointer" onClick={this.uploadFile} title={Messages.get("label.title.changeProfileImage")}/>
-  												</div> : "")}
-
-  												<div>
-  													<form>
-  														<div className="form-group form-profile">
-  															<label className = "fpdi-text-label"> {Messages.getEditable("label.name","fpdi-nav-label")} </label>
-  																<p id = "p-profileUser"> {(this.state.model.name.length>25)?(string(this.state.model.name).trim().substr(0,25).concat("...").toString()):(this.state.model.name)} </p>
-
-  															<label className = "fpdi-text-label"> {Messages.getEditable("label.email","fpdi-nav-label")} </label>
-    															<p id = "p-profileUser"> {this.state.model.email} </p>
-
-    														<label className = "fpdi-text-label"> {Messages.getEditable("label.cpf","fpdi-nav-label")} </label>
-    															<p id = "p-profileUser"> {this.state.model.cpf} </p>
-
-    														<label className = "fpdi-text-label"> {Messages.getEditable("label.birthdate","fpdi-nav-label")} </label>
-    															<p id = "p-profileUser"> {this.state.model.birthdate == null ? this.state.model.birthdate : this.state.model.birthdate.split(" ")[0]} </p>
-
-    														<label className = "fpdi-text-label">{Messages.getEditable("label.cellphone","fpdi-nav-label")}</label>
-    															<p id = "p-profileUser"> {this.state.model.cellphone} </p>
-
-															<label className = "fpdi-text-label">{Messages.getEditable("label.phone","fpdi-nav-label")}</label>
-    															<p id = "p-profileUser"> {this.state.model.phone}  </p>
-
-    														<label className = "fpdi-text-label"> {Messages.getEditable("label.department","fpdi-nav-label")}</label>
-    															<p id = "p-profileUser"> {this.state.model.department == null ? this.state.model.department: (this.state.model.department.length>25)?(string(this.state.model.department).trim().substr(0,25).concat("...").toString()):(this.state.model.department)} </p>
-
-    														<label className = "fpdi-text-label"> {Messages.getEditable("label.userType","fpdi-nav-label")} </label>
-    															<p id = "p-profileUser"> {AccessLevels.mapped[this.state.model.accessLevel]} </p>
-  														</div>
-  													</form>
-  												</div>
-
-      										</div>
-   							 			</div>
-
-   							 		:
-
-   							 		<div className="panel panel-default panel-default-user">
-      									<div className="panel-heading">{Messages.getEditable("label.userData","fpdi-nav-label")}</div>
-      										<div className="panel-body">
-
-      											<div className="fpdi-container-profile">
-													<img className="fpdi-image-profile" src={this.state.model.picture == null ?(Logo):this.state.model.picture}/>
-												</div>
-												{(this.context.roles.ADMIN  || _.contains(this.context.permissions,
-														"org.forpdi.core.user.authz.permission.ManageUsersPermission") ?
-												<div className="fpdi-container-profile-icon">
-  													<span className="mdi mdi-camera mdi-camera-attributes cursorPointer" onClick={this.uploadFile} title={Messages.get("label.title.changeProfileImage")}/>
-  												</div> : "")}
-
-  												<div className="panel-body">
-   							 							<VerticalForm
-				    										ref="editUser"
-															onSubmit={this.onSubmit}
-															fields={this.state.fields}
-															store={UserStore}
-															submitLabel={Messages.get("label.submitLabel")}
-															onCancel = {this.onCancel}
-														/>
-
-												</div>
-
-      										</div>
-
-
-   							 		</div> }
-
-							</div>
-
-								<div className="col-sm-3">
-									<div className="panel panel-default panel-default-user">
-      									<div className="panel-heading"> {Messages.getEditable("label.userPermissions","fpdi-nav-label")}
-      										{(this.state.editingPermission || !(this.context.roles.ADMIN  || _.contains(this.context.permissions,
-														"org.forpdi.core.user.authz.permission.ManageUsersPermission")) ? "" :
-      											<span className="mdi mdi-pencil cursorPointer floatRight" onClick={this.editPermissions} title={Messages.get("label.title.editPermissions")}/>)}
-      									</div>
-      									<div className="padding5">
-											{this.state.permissions.map((item, idx) => {
-												if(item.granted || this.state.model.accessLevel >= item.accessLevel ||
-													(this.state.editingPermission && this.context.accessLevel >= item.accessLevel)){
-													return(
-														<div key={"perm-"+idx} className="user-permission-list-item">
-															{(this.state.editingPermission ?
-																<input type="checkbox" defaultChecked={item.granted || this.state.model.accessLevel >= item.accessLevel}
-															 		className="user-permission-checkbox" disabled={this.state.model.accessLevel >= item.accessLevel ||
-															 		this.context.accessLevel < item.accessLevel} onClick={this.selectPermission.bind(this, idx)}/> : ""
-															 	)
-															}
-
-															{item.permission}
-															<span className={"mdi cursorPointer floatRight "+(_.contains(this.state.detailed, idx) ?
-															 "mdi-chevron-down" : "mdi-chevron-right")} onClick={this.details.bind(this, idx)}/>
-															{(_.contains(this.state.detailed, idx) ? this.showDescription(item.description) : "")}
-														</div>
-													);
-												}
-											})}
-											{!!this.state.editingPermission ?
-												(<div className="form-group user-permission-btn-ctn">
-													<button className="btn btn-success user-permission-btn-save" onClick={this.savePermissions}>{Messages.getEditable("label.submitLabel","fpdi-nav-label")}</button>
-													<button className="btn btn-default" onClick={this.editPermissions}>{Messages.getEditable("label.cancel","fpdi-nav-label")}</button>
-												</div>)
-												:
-												""
-											}
+						<div className="panel panel-default panel-message">
+							<div className="panel-heading"> {Messages.getEditable("label.messageHistory", "fpdi-nav-label")} </div>
+							<div className="padding5">
+								<div className="panel-body">
+									<NotificationMessageUser ref="pagination" />
+									{this.state.hideShowMoreMessages == false ?
+										<div className="textAlignCenter">
+											<a onClick={this.showMoreMessages}>{Messages.getEditable("label.viewMore", "fpdi-nav-label")}</a>
 										</div>
-      								</div>
-								</div>
-								<div className="col-sm-6">
-									<div className="panel panel-default panel-default-user">
-										<div className="panel-heading">{Messages.getEditable("label.sendMessages","fpdi-nav-label")}</div>
-										<div className="padding5">
-											<div className="panel-body">
-												<VerticalForm
-				    								ref="messageUser"
-													onSubmit={this.onSubmitMenssage}
-													fields={this.state.fieldsMessage}
-													store={UserStore}
-													submitLabel={Messages.get("label.send")}
-													cancelLabel={Messages.get("label.clean")}
-													onCancel = {this.onClear}
-												/>
-											</div>
-										</div>
-									</div>
-
-									<div className = "panel panel-default panel-message">
-										<div className="panel-heading"> {Messages.getEditable("label.messageHistory","fpdi-nav-label")} </div>
-										<div className="padding5">
-											<div className="panel-body">
-												<NotificationMessageUser ref="pagination"/>
-        										{this.state.hideShowMoreMessages == false ?
-        											<div className="textAlignCenter">
-                    									<a onClick={this.showMoreMessages}>{Messages.getEditable("label.viewMore","fpdi-nav-label")}</a>
-                									</div>
-                								:""}
-											</div>
-										</div>
-									</div>
-
+										: ""}
 								</div>
 							</div>
-					</div>);
+						</div>
+
+					</div>
+				</div>
+			</div>);
 
 	}
 
