@@ -33,6 +33,11 @@ export default React.createClass({
 				fieldId: this.props.fieldDef.id
 			};
 		}
+
+		if(this.props.fieldDef.fieldName){
+			this.props.fieldDef.name=this.props.fieldDef.fieldName
+		}
+
 		return {
 			fieldId: "field-"+this.props.fieldDef.name.replace(/\./g, "-")
 		};
@@ -151,6 +156,7 @@ export default React.createClass({
 					placeholder={this.props.fieldDef.placeholder}
 					name={this.props.fieldDef.name}
 					defaultValue={this.props.fieldDef.value}
+					value={this.props.fieldDef.currValue}
 					maxLength={this.props.fieldDef.maxLength*1.01}
 					id={this.state.fieldId}
 					ref={this.state.fieldId}
@@ -161,10 +167,11 @@ export default React.createClass({
 		} else if (this.props.fieldDef.type == 'select') {
 			fieldEl = (
 				<select
-					className="form-control"
+					className= {"form-control "+this.props.fieldDef.className}
 					placeholder={this.props.fieldDef.placeholder}
 					name={this.props.fieldDef.name}
 					defaultValue={this.props.fieldDef.value}
+					value={this.props.fieldDef.currValue}
 					id={this.state.fieldId}
 					ref={this.state.fieldId}
 					onChange={this.props.fieldDef.onChange || _.noop}
@@ -172,13 +179,20 @@ export default React.createClass({
 						{string(this.props.fieldDef.placeholder).isEmpty() ? "":
 							<option value="" data-placement="right" title={this.props.fieldDef.placeholder}>{this.props.fieldDef.placeholder}</option>
 						}
+
 						{this.props.fieldDef.options ? this.props.fieldDef.options.map((opt,idx) => {
 							return (<option key={'field-opt-'+this.state.fieldId+"-"+idx} value={opt.get ? opt.get(this.props.fieldDef.valueField):opt[this.props.fieldDef.valueField]}
 								data-placement="right" title={opt[this.props.fieldDef.displayField]}>
-									{ this.props.fieldDef.renderDisplay ?
-										this.props.fieldDef.renderDisplay(opt)
-										:
-										(opt.get ?string(opt.get(this.props.fieldDef.displayField)).trim().truncate(80, "...").toString() :opt[this.props.fieldDef.displayField])
+									{ this.props.fieldDef.renderDisplay
+									? this.props.fieldDef.renderDisplay(opt)
+									:(
+										(opt.get ?string(opt.get(this.props.fieldDef.displayField)).trim().truncate(80, "...").toString()
+											:((opt[this.props.fieldDef.displayField].length > this.props.limit)
+												? opt[this.props.fieldDef.displayField].substring(0,this.props.limit-3)+"..."
+												: opt[this.props.fieldDef.displayField]
+											)
+										)
+									)
 									}
 							</option>);
 						}):''}
@@ -196,6 +210,7 @@ export default React.createClass({
 						id={this.state.fieldId+"-"+idx}
 						defaultChecked={(this.props.fieldDef.value == null)?(idx==0?"checked":null):(checked ? "checked":null)}
 						value={fieldVal}
+						onClick={this.props.fieldDef.onClick || _.noop}
 						/> { this.props.fieldDef.renderDisplay ?
 							this.props.fieldDef.renderDisplay(opt)
 							:
@@ -218,6 +233,22 @@ export default React.createClass({
 					placeholderText="DD/MM/AAAA"
 					showYearDropdown
 					/></div>);
+		} else if (this.props.fieldDef.type == 'custom-mask') {
+			fieldEl = (<MaskedInput
+				mask={this.props.fieldDef.mask}
+				type="tel"
+				className="form-control"
+				name={this.props.fieldDef.name}
+				value={this.props.fieldDef.value}
+				id={this.state.fieldId}
+				ref={this.state.fieldId}
+				placeholder={this.props.fieldDef.placeholder}
+				onChange={this.props.fieldDef.onChange || _.noop}
+				// placeholder=""
+				onKeyPress={this.onKeyUp}
+				onPaste={this.onKeyUp}
+			/>);
+
 		} else if (this.props.fieldDef.type == 'daterange') {
 			fieldEl = (<input
 				className="form-control"
@@ -355,11 +386,12 @@ export default React.createClass({
 				/>);
 			} else {
 				fieldEl = (<input
-					className="form-control"
+					className={"form-control "+ this.props.className}
 					type={this.props.fieldDef.type}
 					name={this.props.fieldDef.name}
 					maxLength={this.props.fieldDef.maxLength}
 					defaultValue={this.props.fieldDef.value}
+					value={this.props.fieldDef.currValue}
 					id={this.state.fieldId}
 					ref={this.state.fieldId}
 					placeholder={this.props.fieldDef.placeholder}
