@@ -1,5 +1,5 @@
 import React from "react";
-import {Link} from 'react-router';
+import { Link } from 'react-router';
 
 import TreeView from "forpdi/jsx/core/widget/treeview/TreeView.jsx";
 import StructureStore from "forpdi/jsx/planning/store/Structure.jsx";
@@ -63,11 +63,11 @@ export default React.createClass({
 
 	getFields(model, agg) {
 		var fields = [];
-		if(model.data.level.indicator){
+		if (model.data.level.indicator) {
 			var options = [{
-				name:  Messages.get("label.aggregateSimples"),
+				name: Messages.get("label.aggregateSimples"),
 				value: !model.data.aggregate
-			},{
+			}, {
 				name: Messages.get("label.aggregateAggregated"),
 				value: model.data.aggregate
 			}];
@@ -84,7 +84,7 @@ export default React.createClass({
 				onClick: this.onIndicatorTypeClick,
 				extraRender: this.renderAggregate,
 				disabled: this.context.roles.MANAGER || _.contains(this.context.permissions,
-         							PermissionsTypes.MANAGE_PLAN_PERMISSION) ? false : true
+					PermissionsTypes.MANAGE_PLAN_PERMISSION) ? false : true
 			});
 		}
 
@@ -93,20 +93,21 @@ export default React.createClass({
 			type: "text",
 			required: true,
 			placeholder: "",
-			label: Messages.getEditable("label.name","fpdi-nav-label"),
+			label: Messages.getEditable("label.name", "fpdi-nav-label"),
 			value: model.data.name,
 			disabled: this.context.roles.MANAGER || _.contains(this.context.permissions,
-         							PermissionsTypes.MANAGE_PLAN_PERMISSION) ? false : true
+				PermissionsTypes.MANAGE_PLAN_PERMISSION) ? false : true
 		});
 
 		if (model.data.level.attributes) {
-			model.data.level.attributes.map((model,idx) =>{
+			model.data.level.attributes.map((model, idx) => {
+
 				var formattedValue = (model.attributeInstance ? model.attributeInstance.formattedValue : "");
 				var value = (model.attributeInstance ? model.attributeInstance.value : "");
 				var editModeValue = null;
 				if(model.type == AttributeTypes.NUMBER_FIELD && value != undefined) {
 					if (model.attributeInstance){  //&& model.attributeInstance.valueAsNumber) {
-						value = model.attributeInstance.valueAsNumber;
+						value = model.attributeInstance.valueAsNumber || '';
 						editModeValue = value.toString().replace(".", ",");
 					} else {
 						value = "";
@@ -115,6 +116,9 @@ export default React.createClass({
 
 				}
 
+				if (model.type == AttributeTypes.RESPONSIBLE_FIELD && model.attributeInstance && model.attributeInstance.valueAsNumber) {
+					value = model.attributeInstance.valueAsNumber
+				}
 
 				if (model.beginField) {
 					dateBegin = value;
@@ -134,7 +138,7 @@ export default React.createClass({
 
 				if (model.formatField && agg) {
 					fields.push({
-						name: "attribute"+idx,
+						name: "attribute" + idx,
 						type: model.type,
 						required: model.required,
 						placeholder: model.description,
@@ -148,7 +152,7 @@ export default React.createClass({
 						schedule: model.schedule,
 						optionsField: model.optionsField,
 						users: model.users,
-						onChange:(model.beginField ? this.onChangeBegin : model.endField ? this.onChangeEnd : model.finishDate ? this.onChangeFinish : null) ,
+						onChange: (model.beginField ? this.onChangeBegin : model.endField ? this.onChangeEnd : model.finishDate ? this.onChangeFinish : null),
 						extraSchedule: (model.schedule ? this.saveInstanceSchedule : undefined),
 						actionPlans: model.actionPlans,
 						extraActionPlans: (model.actionPlans ? this.newLevelInstanceActionPlan : undefined),
@@ -157,7 +161,7 @@ export default React.createClass({
 					});
 				} else if (model.reachedField || model.justificationField) {
 					fields.push({
-						name: "attribute"+idx,
+						name: "attribute" + idx,
 						type: model.type,
 						required: model.required,
 						placeholder: model.description,
@@ -171,7 +175,7 @@ export default React.createClass({
 						schedule: model.schedule,
 						optionsField: model.optionsField,
 						users: model.users,
-						onChange:(model.beginField ? this.onChangeBegin : model.endField ? this.onChangeEnd : model.finishDate ? this.onChangeFinish : null) ,
+						onChange: (model.beginField ? this.onChangeBegin : model.endField ? this.onChangeEnd : model.finishDate ? this.onChangeFinish : null),
 						extraSchedule: (model.schedule ? this.saveInstanceSchedule : undefined),
 						actionPlans: model.actionPlans,
 						extraActionPlans: (model.actionPlans ? this.newLevelInstanceActionPlan : undefined),
@@ -180,7 +184,7 @@ export default React.createClass({
 					});
 				} else {
 					fields.push({
-						name: "attribute"+idx,
+						name: "attribute" + idx,
 						type: model.type,
 						required: model.required,
 						placeholder: model.description,
@@ -189,13 +193,13 @@ export default React.createClass({
 						formattedValue: formattedValue,
 						editModeValue: editModeValue,
 						disabled: this.context.roles.MANAGER || _.contains(this.context.permissions,
-         							PermissionsTypes.MANAGE_PLAN_PERMISSION) ? false : true,
+							PermissionsTypes.MANAGE_PLAN_PERMISSION) ? false : true,
 						budgets: model.budgets,
 						extra: (model.budgets ? this.newLevelInstanceBudget : undefined),
 						schedule: model.schedule,
 						optionsField: model.optionsField,
 						users: model.users,
-						onChange:(model.beginField ? this.onChangeBegin : model.endField ? this.onChangeEnd : model.finishDate ? this.onChangeFinish : null) ,
+						onChange: (model.beginField ? this.onChangeBegin : model.endField ? this.onChangeEnd : model.finishDate ? this.onChangeFinish : null),
 						extraSchedule: (model.schedule ? this.saveInstanceSchedule : undefined),
 						actionPlans: model.actionPlans,
 						extraActionPlans: (model.actionPlans ? this.newLevelInstanceActionPlan : undefined),
@@ -215,24 +219,24 @@ export default React.createClass({
 		return fields;
 	},
 
-	onIndicatorTypeClick(evt){
+	onIndicatorTypeClick(evt) {
 		var idx = evt.target.id.split('-')[1];
 		let normalRadio = this.refs['levelForm'].refs['indicator-type'].refs['indicator-type-option-0'];
 		let aggregateRadio = this.refs['levelForm'].refs['indicator-type'].refs['indicator-type-option-1'];
 
 		let agg;
-		if(idx == 0){
+		if (idx == 0) {
 			aggregateRadio.checked = false;
 			aggregateRadio.value = false;
 			normalRadio.checked = true;
 			normalRadio.value = true;
 			agg = false;
-		}else{
-			if(this.state.haveChildren){
+		} else {
+			if (this.state.haveChildren) {
 				this.context.toastr.addAlertError(Messages.get("label.error.noGoalsAggregateAggregated"));
-        		aggregateRadio.checked = false;
+				aggregateRadio.checked = false;
 				aggregateRadio.value = false;
-        		return false;
+				return false;
 			}
 			aggregateRadio.checked = true;
 			aggregateRadio.value = true;
@@ -253,7 +257,7 @@ export default React.createClass({
 
 		StructureStore.on("levelGoalClosed", (model) => {
 			this.setState({
-				vizualization:true
+				vizualization: true
 			})
 		});
 
@@ -266,7 +270,7 @@ export default React.createClass({
 		});
 
 		StructureStore.on('levelAttributeSaved', (model) => {
-			if(model.success == true){
+			if (model.success == true) {
 				this.context.toastr.addAlertSuccess(Messages.get("label.success.informationSaved"));
 				this.refs.levelInstanceTitle.title = model.data.name;
 				var showPolarityAlert = false;
@@ -283,7 +287,7 @@ export default React.createClass({
 					showPolarityAlert: showPolarityAlert
 				});
 			}
-			else{
+			else {
 				this.context.toastr.addAlertError(Messages.get("label.impossibleSaveAttributes"));
 			}
 		}, me);
@@ -300,7 +304,7 @@ export default React.createClass({
 		StructureStore.on('favoriteSaved', model => {
 			this.setState({
 				favoriteExistent: true,
-				favoriteTotal: this.state.favoriteTotal+1
+				favoriteTotal: this.state.favoriteTotal + 1
 			});
 			this.context.toastr.addAlertSuccess(model.data.levelInstance.name + " " + Messages.get("label.addedFavorites"));
 		}, me);
@@ -308,7 +312,7 @@ export default React.createClass({
 		StructureStore.on('favoriteRemoved', model => {
 			this.setState({
 				favoriteExistent: false,
-				favoriteTotal: this.state.favoriteTotal-1
+				favoriteTotal: this.state.favoriteTotal - 1
 			});
 			this.context.toastr.addAlertSuccess(model.data.levelInstance.name + " " + Messages.get("label.removedFavorites"));
 		}, me);
@@ -332,12 +336,12 @@ export default React.createClass({
 
 		StructureStore.on("levelAttributeNoRetrieved", (model) => {
 			var url = window.location.href;
-			window.location.assign(url.split("#")[0]+"#/home");
+			window.location.assign(url.split("#")[0] + "#/home");
 		}, me);
 
 		StructureStore.on("levelSonsRetrieved", (model) => {
 			me.setState({
-				undeletable: model.data.sons.list.length>0 ? false : true
+				undeletable: model.data.sons.list.length > 0 ? false : true
 			});
 		}, me);
 	},
@@ -346,7 +350,7 @@ export default React.createClass({
 		StructureStore.off(null, null, this);
 		BudgetStore.off(null, null, this);
 		ScheduleStore.off(null, null, this);
- 		ActionPlanStore.off(null, null, this);
+		ActionPlanStore.off(null, null, this);
 	},
 
 	componentWillReceiveProps(newProps) {
@@ -364,7 +368,7 @@ export default React.createClass({
 		}
 	},
 
-	getLevelInstanceAttributes(levelInstanceId){
+	getLevelInstanceAttributes(levelInstanceId) {
 		StructureStore.dispatch({
 			action: StructureStore.ACTION_RETRIEVE_LEVELATTRIBUTES,
 			data: {
@@ -373,34 +377,34 @@ export default React.createClass({
 		});
 	},
 
-	newLevelInstanceBudget(subAction,name,committed,realized) {
+	newLevelInstanceBudget(subAction, name, committed, realized) {
 		BudgetStore.dispatch({
 			action: BudgetStore.ACTION_CREATE,
 			data: {
 				subAction: subAction,
 				name: name,
-				committed:committed,
-				realized:realized,
+				committed: committed,
+				realized: realized,
 				instanceId: this.props.params.levelInstanceId,
 
 			}
 		});
 	},
 
-	saveInstanceSchedule(scheduleInstance){
+	saveInstanceSchedule(scheduleInstance) {
 		ScheduleStore.dispatch({
 			action: ScheduleStore.ACTION_CUSTOM_SAVE,
 			data: {
 				scheduleInstance: {
- 					id: scheduleInstance.id,
- 					number: scheduleInstance.number,
- 					description: scheduleInstance.description,
- 					periodicity: scheduleInstance.periodicity,
- 					scheduleValues: scheduleInstance.scheduleValues
+					id: scheduleInstance.id,
+					number: scheduleInstance.number,
+					description: scheduleInstance.description,
+					periodicity: scheduleInstance.periodicity,
+					scheduleValues: scheduleInstance.scheduleValues
 				},
 				scheduleId: scheduleInstance.scheduleId,
- 				beginDate: scheduleInstance.begin,
- 				endDate: scheduleInstance.end
+				beginDate: scheduleInstance.begin,
+				endDate: scheduleInstance.end
 			},
 			opts: {
 				wait: true
@@ -408,7 +412,7 @@ export default React.createClass({
 		});
 	},
 
-	saveInstanceTable(tableInstance){
+	saveInstanceTable(tableInstance) {
 		TableStore.dispatch({
 			action: TableStore.ACTION_CUSTOM_SAVE,
 			data: {
@@ -424,7 +428,7 @@ export default React.createClass({
 		});
 	},
 
-	newLevelInstanceActionPlan(actionPlan){
+	newLevelInstanceActionPlan(actionPlan) {
 		ActionPlanStore.dispatch({
 			action: ActionPlanStore.ACTION_CUSTOM_CREATE,
 			data: {
@@ -445,40 +449,42 @@ export default React.createClass({
 	},
 
 
-	confirmCompleteGoal(id,closeOpenGoal) {
+	confirmCompleteGoal(id, closeOpenGoal) {
 		var msg = "";
 
-		{closeOpenGoal ?
-			msg = Messages.get("label.msg.completeGoal")
+		{
+			closeOpenGoal ?
+				msg = Messages.get("label.msg.completeGoal")
 
-			: msg = Messages.get("label.msg.openGoal")
+				: msg = Messages.get("label.msg.openGoal")
 		}
 
 		Modal.confirmCustom(() => {
 			Modal.hide();
-			{StructureStore.dispatch({
-				action: StructureStore.ACTION_CLOSE_GOAL,
-				data: {
-					id: id,
-					openCloseGoal:closeOpenGoal
-				}
-			});
+			{
+				StructureStore.dispatch({
+					action: StructureStore.ACTION_CLOSE_GOAL,
+					data: {
+						id: id,
+						openCloseGoal: closeOpenGoal
+					}
+				});
 				var newModel = this.state.model;
 				newModel.data.closed = !newModel.data.closed;
 				this.setState({
 					model: newModel
 				})
 			};
-		},msg,
-		() => {
-			Modal.hide();
-		});
+		}, msg,
+			() => {
+				Modal.hide();
+			});
 	},
 
 	deleteLevelAttribute() {
 		var data = this.state.model;
-		if(this.state.undeletable) {
-			var msg =  Messages.get("label.deleteConfirmation") + " "  + this.state.model.data.name + "?";
+		if (this.state.undeletable) {
+			var msg = Messages.get("label.deleteConfirmation") + " " + this.state.model.data.name + "?";
 			Modal.confirmCancelCustom(() => {
 				Modal.hide();
 				var levelInstance = {
@@ -489,7 +495,7 @@ export default React.createClass({
 					data: levelInstance
 				});
 
-			},msg,()=>{Modal.hide()});
+			}, msg, () => { Modal.hide() });
 
 			/*Modal.deleteConfirmCustom(() => {
 				Modal.hide();
@@ -505,7 +511,7 @@ export default React.createClass({
 
 	},
 
-	saveFavoriteLevel(){
+	saveFavoriteLevel() {
 		StructureStore.dispatch({
 			action: StructureStore.ACTION_SAVE_FAVORITE,
 			data: {
@@ -514,24 +520,24 @@ export default React.createClass({
 		});
 	},
 
-	removeFavoriteLevel(){
-		var msg =  Messages.get("label.msg.removeConfirmation")  + " " + this.state.model.data.name + " " + Messages.get("label.msg.favorites");;
-			Modal.confirmCancelCustom(() => {
-				Modal.hide();
+	removeFavoriteLevel() {
+		var msg = Messages.get("label.msg.removeConfirmation") + " " + this.state.model.data.name + " " + Messages.get("label.msg.favorites");;
+		Modal.confirmCancelCustom(() => {
+			Modal.hide();
 
-				StructureStore.dispatch({
-					action: StructureStore.ACTION_REMOVE_FAVORITE,
-					data: {
-						levelInstanceId: this.state.model.data.id
-					}
-				});
-			},msg,()=>{Modal.hide()}
+			StructureStore.dispatch({
+				action: StructureStore.ACTION_REMOVE_FAVORITE,
+				data: {
+					levelInstanceId: this.state.model.data.id
+				}
+			});
+		}, msg, () => { Modal.hide() }
 		);
 	},
 
 	isNumber(n) {
 
-    	return !isNaN(parseFloat(n)) && isFinite(n);
+		return !isNaN(parseFloat(n)) && isFinite(n);
 	},
 
 	onSubmit(data) {
@@ -541,11 +547,11 @@ export default React.createClass({
 
 		if (validation.boolMsg) {
 			this.context.toastr.addAlertError(validation.msg);
- 			return;
+			return;
 		}
 
 		var aggregateArray = [];
-		if(this.state.model.data.level.indicator && this.state.aggregate){
+		if (this.state.model.data.level.indicator && this.state.aggregate) {
 			var indicators = this.refs['levelForm'].refs['indicator-type'].refs['agg-ind-config'].selectedIndicators;
 			indicators.map((indicator, idx) => {
 				aggregateArray.push({
@@ -588,7 +594,7 @@ export default React.createClass({
 		});
 	},
 	onCancel() {
-		if(this.state.aggregate == true) {
+		if (this.state.aggregate == true) {
 			this.setState({
 				vizualization: !this.state.vizualization,
 				aggregate: this.state.model.data.aggregate
@@ -619,37 +625,37 @@ export default React.createClass({
 		});
 	},
 
-	onChangeBegin(data){
+	onChangeBegin(data) {
 		var model = this.state.model;
 		this.setState({
-			model:model,
+			model: model,
 		});
-		this.refs['levelForm'].refs["attribute"+this.state.beginIdx].props.fieldDef.value = data.format('DD/MM/YYYY');
+		this.refs['levelForm'].refs["attribute" + this.state.beginIdx].props.fieldDef.value = data.format('DD/MM/YYYY');
 	},
-	onChangeEnd(data){
+	onChangeEnd(data) {
 		var model = this.state.model;
 		this.setState({
-			model:model,
+			model: model,
 		});
-		this.refs['levelForm'].refs["attribute"+this.state.endIdx].props.fieldDef.value = data.format('DD/MM/YYYY');
+		this.refs['levelForm'].refs["attribute" + this.state.endIdx].props.fieldDef.value = data.format('DD/MM/YYYY');
 	},
-	onChangeFinish(data){
+	onChangeFinish(data) {
 		var model = this.state.model;
 		this.setState({
-			model:model,
+			model: model,
 		});
-		this.refs['levelForm'].refs["attribute"+this.state.finishIdx].props.fieldDef.value = data.format('DD/MM/YYYY');
+		this.refs['levelForm'].refs["attribute" + this.state.finishIdx].props.fieldDef.value = data.format('DD/MM/YYYY');
 	},
 
-	onlyNumber(evt){
- 		var key = evt.which;
- 		if((key < 48 || key > 57)) {
- 			evt.preventDefault();
- 			return;
- 		}
- 	},
+	onlyNumber(evt) {
+		var key = evt.which;
+		if ((key < 48 || key > 57)) {
+			evt.preventDefault();
+			return;
+		}
+	},
 
-	isParent(model){
+	isParent(model) {
 		this.setState({
 			haveChildren: model.data.sons.list.length > 0
 		});
@@ -663,45 +669,45 @@ export default React.createClass({
 		var w = window.open(url);
 	},
 
-	renderAggregate(){
-		if(this.state.aggregate){
-			return(<AggregateIndicator
+	renderAggregate() {
+		if (this.state.aggregate) {
+			return (<AggregateIndicator
 				ref="agg-ind-config"
 				indicatorsList={this.state.model.data.indicatorList}
 				calculationType={this.state.model.data.calculation}
 				visualization={this.state.vizualization}
-				selfId={this.props.params.levelInstanceId}/>);
+				selfId={this.props.params.levelInstanceId} />);
 		}
 	},
 
 	renderBreadcrumb() {
-		return(
+		return (
 			<div>
 				<span>
 					<Link className="fpdi-breadcrumb fpdi-breadcrumbDivisor"
-						to={'/plan/'+this.context.planMacro.attributes.id+'/details/'}
-						title={this.context.planMacro.attributes.name}>{this.context.planMacro.attributes.name.length > 15 ? this.context.planMacro.attributes.name.substring(0, 15)+"..." : this.context.planMacro.attributes.name.substring(0, 15)
-					}</Link>
+						to={'/plan/' + this.context.planMacro.attributes.id + '/details/'}
+						title={this.context.planMacro.attributes.name}>{this.context.planMacro.attributes.name.length > 15 ? this.context.planMacro.attributes.name.substring(0, 15) + "..." : this.context.planMacro.attributes.name.substring(0, 15)
+						}</Link>
 					<span className="mdi mdi-chevron-right fpdi-breadcrumbDivisor"></span>
 				</span>
 				<span>
 					<Link className="fpdi-breadcrumb fpdi-breadcrumbDivisor"
-						to={'/plan/'+this.context.planMacro.attributes.id+'/details/subplan/'+this.state.model.data.plan.id}
-						title={this.state.model.data.plan.name}>{this.state.model.data.plan.name.length > 15 ? this.state.model.data.plan.name.substring(0, 15)+"..." : this.state.model.data.plan.name.substring(0, 15)}</Link>
+						to={'/plan/' + this.context.planMacro.attributes.id + '/details/subplan/' + this.state.model.data.plan.id}
+						title={this.state.model.data.plan.name}>{this.state.model.data.plan.name.length > 15 ? this.state.model.data.plan.name.substring(0, 15) + "..." : this.state.model.data.plan.name.substring(0, 15)}</Link>
 					<span className="mdi mdi-chevron-right fpdi-breadcrumbDivisor"></span>
 				</span>
 				{this.state.model.data.parents.map((parent, idx) => {
 					return (
 						<span key={idx}>
 							<Link className="fpdi-breadcrumb fpdi-breadcrumbDivisor"
-								to={"/plan/"+this.context.planMacro.attributes.id+"/details/subplan/level/"+parent.id}
-								title={parent.name}>{parent.name.length > 15 ? parent.name.substring(0, 15)+"..." : parent.name.substring(0, 15)}</Link>
+								to={"/plan/" + this.context.planMacro.attributes.id + "/details/subplan/level/" + parent.id}
+								title={parent.name}>{parent.name.length > 15 ? parent.name.substring(0, 15) + "..." : parent.name.substring(0, 15)}</Link>
 							<span className="mdi mdi-chevron-right fpdi-breadcrumbDivisor"></span>
 						</span>
 					);
 				})}
 				<span className="fpdi-breadcrumb fpdi-selectedOnBreadcrumb">
-					{this.state.model.data.name.length > 15 ? this.state.model.data.name.substring(0, 15)+"..." : this.state.model.data.name.substring(0, 15)}
+					{this.state.model.data.name.length > 15 ? this.state.model.data.name.substring(0, 15) + "..." : this.state.model.data.name.substring(0, 15)}
 				</span>
 			</div>
 		);
@@ -714,22 +720,22 @@ export default React.createClass({
 					<a
 						className="mdi mdi-pencil disabledIcon"
 						title={Messages.get("label.title.UnableEditArchivedPlan")}>
-						<span id = "menu-levels">
-							{Messages.getEditable("label.title.UnableEditArchivedPlan","fpdi-nav-label")}
+						<span id="menu-levels">
+							{Messages.getEditable("label.title.UnableEditArchivedPlan", "fpdi-nav-label")}
 						</span>
 					</a>
-		         </li>
+				</li>
 
-		         <li>
+				<li>
 					<a
 						className="mdi mdi-delete disabledIcon"
 						title={Messages.get("label.unableDeleteArchivedPlan")}
 						type="submit">
-						<span id = "menu-levels">
-							{Messages.getEditable("label.unableDeleteArchivedPlan","fpdi-nav-label")}
+						<span id="menu-levels">
+							{Messages.getEditable("label.unableDeleteArchivedPlan", "fpdi-nav-label")}
 						</span>
 					</a>
-		         </li>
+				</li>
 			</ul>
 		);
 	},
@@ -738,182 +744,182 @@ export default React.createClass({
 		return (
 			<ul id="level-menu" className="dropdown-menu" >
 				{((this.context.roles.MANAGER || _.contains(this.context.permissions,
- 							PermissionsTypes.MANAGE_PLAN_PERMISSION)) && this.state.vizualization) ||
- 							(userResponsible && userResponsible.id == UserSession.get("user").id &&
- 							this.state.model.data.level.goal) ? (
-					<li>
-						<a onClick={this.editingAttributes}>
-							<span className="mdi mdi-pencil cursorPointer cursorPointer" title={Messages.get("label.title.editInformation")}>
-								<span id="menu-levels"> {Messages.getEditable("label.title.editInformation","fpdi-nav-label")} </span>
-							</span>
-						</a>
+					PermissionsTypes.MANAGE_PLAN_PERMISSION)) && this.state.vizualization) ||
+					(userResponsible && userResponsible.id == UserSession.get("user").id &&
+						this.state.model.data.level.goal) ? (
+						<li>
+							<a onClick={this.editingAttributes}>
+								<span className="mdi mdi-pencil cursorPointer cursorPointer" title={Messages.get("label.title.editInformation")}>
+									<span id="menu-levels"> {Messages.getEditable("label.title.editInformation", "fpdi-nav-label")} </span>
+								</span>
+							</a>
 
-			        </li>) : ""}
+						</li>) : ""}
 
 				{(this.context.roles.MANAGER || _.contains(this.context.permissions,
-	         					PermissionsTypes.MANAGE_PLAN_PERMISSION)) && this.state.model.data.level.name == "Meta" ? (
-					<li>
-						<a onClick={this.state.model.data.closed == false ?
-								this.confirmCompleteGoal.bind(this,this.state.model.data.id, true) : ((this.context.roles.MANAGER || _.contains(this.context.permissions,
-	         					PermissionsTypes.MANAGE_PLAN_PERMISSION)) ? this.confirmCompleteGoal.bind(this, this.state.model.data.id, false) : "") }>
-							<span className={this.state.model.data.closed == false ? "mdi mdi-lock-open-outline lockGoal-open deleteIcon":"mdi  mdi-lock lockGoal-closedeleteIcon"}
-								title= {this.state.model.data.closed == false ? Messages.get("label.title.finishGoal") : ((this.context.roles.MANAGER || _.contains(this.context.permissions,
-	         					PermissionsTypes.MANAGE_PLAN_PERMISSION)) ? "Abrir Meta" : Messages.get("label.title.reopenGoal"))}>
-								<span id="menu-levels"> {this.state.model.data.closed == false ? Messages.get("label.msg.finishGoal"):Messages.get("label.msg.reopenGoal")} </span>
-							</span>
-						</a>
-					</li>) : ""}
-         		{(this.context.roles.MANAGER || _.contains(this.context.permissions,
-         							PermissionsTypes.MANAGE_PLAN_PERMISSION)) && this.state.vizualization ? (
-			        <li>
-						<a onClick={this.exportLevelAttributes}>
-							<span className="mdi mdi-file-export cursorPointer" title={Messages.get("label.generateReport")}>
-								<span id="menu-levels"> {Messages.getEditable("label.generateReport","fpdi-nav-label")} </span>
-							</span>
-						</a>
-					</li>) : ""}
-
-		        {this.context.roles.SYSADMIN ? "" :
-			        (this.state.favoriteExistent ? (
-		         		<li>
-		         			<a onClick={this.removeFavoriteLevel}>
-								<span className="mdi mdi-star-outline cursorPointer" title={Messages.get("label.removeFromFavorites")}>
-									<span id="menu-levels"> {Messages.getEditable("label.removeFromFavorites","fpdi-nav-label")} </span>
+					PermissionsTypes.MANAGE_PLAN_PERMISSION)) && this.state.model.data.level.name == "Meta" ? (
+						<li>
+							<a onClick={this.state.model.data.closed == false ?
+								this.confirmCompleteGoal.bind(this, this.state.model.data.id, true) : ((this.context.roles.MANAGER || _.contains(this.context.permissions,
+									PermissionsTypes.MANAGE_PLAN_PERMISSION)) ? this.confirmCompleteGoal.bind(this, this.state.model.data.id, false) : "")}>
+								<span className={this.state.model.data.closed == false ? "mdi mdi-lock-open-outline lockGoal-open deleteIcon" : "mdi  mdi-lock lockGoal-closedeleteIcon"}
+									title={this.state.model.data.closed == false ? Messages.get("label.title.finishGoal") : ((this.context.roles.MANAGER || _.contains(this.context.permissions,
+										PermissionsTypes.MANAGE_PLAN_PERMISSION)) ? "Abrir Meta" : Messages.get("label.title.reopenGoal"))}>
+									<span id="menu-levels"> {this.state.model.data.closed == false ? Messages.get("label.msg.finishGoal") : Messages.get("label.msg.reopenGoal")} </span>
 								</span>
 							</a>
-			         	</li>)
-			        	: (!this.state.favoriteTotal || this.state.favoriteTotal < 10 ?
-	         		 		<li>
-	         		 			<a onClick={this.saveFavoriteLevel}>
+						</li>) : ""}
+				{(this.context.roles.MANAGER || _.contains(this.context.permissions,
+					PermissionsTypes.MANAGE_PLAN_PERMISSION)) && this.state.vizualization ? (
+						<li>
+							<a onClick={this.exportLevelAttributes}>
+								<span className="mdi mdi-file-export cursorPointer" title={Messages.get("label.generateReport")}>
+									<span id="menu-levels"> {Messages.getEditable("label.generateReport", "fpdi-nav-label")} </span>
+								</span>
+							</a>
+						</li>) : ""}
+
+				{this.context.roles.SYSADMIN ? "" :
+					(this.state.favoriteExistent ? (
+						<li>
+							<a onClick={this.removeFavoriteLevel}>
+								<span className="mdi mdi-star-outline cursorPointer" title={Messages.get("label.removeFromFavorites")}>
+									<span id="menu-levels"> {Messages.getEditable("label.removeFromFavorites", "fpdi-nav-label")} </span>
+								</span>
+							</a>
+						</li>)
+						: (!this.state.favoriteTotal || this.state.favoriteTotal < 10 ?
+							<li>
+								<a onClick={this.saveFavoriteLevel}>
 									<span className="mdi mdi-star cursorPointer" title={Messages.get("label.addFavorites")}>
-										<span id="menu-levels"> {Messages.getEditable("label.addFavorites","fpdi-nav-label")}  </span>
+										<span id="menu-levels"> {Messages.getEditable("label.addFavorites", "fpdi-nav-label")}  </span>
 									</span>
 								</a>
-			         		</li>
-		         			:
-			         			<li>
-					         		<a type="submit">
-										<span className="mdi mdi-star disabledIcon" title={Messages.get("label.maxFavorites")}>
-											<span id="menu-levels"> {Messages.getEditable("label.notAddFavorites","fpdi-nav-label")} </span>
-										</span>
-									</a>
-					         	</li>
-		         		)
-		         	)
-			    }
+							</li>
+							:
+							<li>
+								<a type="submit">
+									<span className="mdi mdi-star disabledIcon" title={Messages.get("label.maxFavorites")}>
+										<span id="menu-levels"> {Messages.getEditable("label.notAddFavorites", "fpdi-nav-label")} </span>
+									</span>
+								</a>
+							</li>
+						)
+					)
+				}
 
-			    {(this.context.roles.MANAGER || _.contains(this.context.permissions,
-         							PermissionsTypes.MANAGE_PLAN_PERMISSION)) && this.state.vizualization ? (
-			        (this.state.undeletable?
-			         	<li>
-							<a onClick={this.deleteLevelAttribute} type="submit">
-								<span id="cacando-a-pala" className="mdi mdi-delete cursorPointer" title={Messages.get("label.delete") + " " + this.state.subTitle}>
-									<span id="menu-levels"> {Messages.get("label.delete")}{/*Messages.getEditable("label.delete","fpdi-nav-label")*/} </span>
-								</span>
-							</a>
-			         	</li>
-			        :
-			        	<li>
-			        		<a type="submit">
-								<span className="mdi mdi-delete disabledIcon marginLeft5 cursorPointer" title={Messages.get("label.notDeleted") +(!this.state.model.data.leaf ? ", possui níveis filhos" : "")}>
-									<span id="menu-levels"> {Messages.getEditable("label.notDeleted","fpdi-nav-label")} </span>
-								</span>
-							</a>
-			         	</li>
-			        )) : ""}
+				{(this.context.roles.MANAGER || _.contains(this.context.permissions,
+					PermissionsTypes.MANAGE_PLAN_PERMISSION)) && this.state.vizualization ? (
+						(this.state.undeletable ?
+							<li>
+								<a onClick={this.deleteLevelAttribute} type="submit">
+									<span id="cacando-a-pala" className="mdi mdi-delete cursorPointer" title={Messages.get("label.delete") + " " + this.state.subTitle}>
+										<span id="menu-levels"> {Messages.get("label.delete")}{/*Messages.getEditable("label.delete","fpdi-nav-label")*/} </span>
+									</span>
+								</a>
+							</li>
+							:
+							<li>
+								<a type="submit">
+									<span className="mdi mdi-delete disabledIcon marginLeft5 cursorPointer" title={Messages.get("label.notDeleted") + (!this.state.model.data.leaf ? ", possui níveis filhos" : "")}>
+										<span id="menu-levels"> {Messages.getEditable("label.notDeleted", "fpdi-nav-label")} </span>
+									</span>
+								</a>
+							</li>
+						)) : ""}
 			</ul>
 		);
 	},
 
 	renderClosedGoal() {
-	 	return (
+		return (
 			<ul id="level-menu" className="dropdown-menu">
 				<li>
 					<a data-placement="top">
 						<span className="mdi mdi-pencil disabledIcon marginLeft5" title={Messages.get("label.title.notEditGoalComplete")}>
-							<span id="menu-levels"> {Messages.getEditable("label.notEdit","fpdi-nav-label")} </span>
+							<span id="menu-levels"> {Messages.getEditable("label.notEdit", "fpdi-nav-label")} </span>
 						</span>
 					</a>
-		         </li>
-
-				 {(this.context.roles.MANAGER || _.contains(this.context.permissions,
-	         					PermissionsTypes.MANAGE_PLAN_PERMISSION)) && this.state.model.data.level.name == "Meta" ? (
-					<li>
-						<a onClick={this.state.model.data.closed == false ?
-								this.confirmCompleteGoal.bind(this,this.state.model.data.id, true) : ((this.context.roles.MANAGER || _.contains(this.context.permissions,
-	         					PermissionsTypes.MANAGE_PLAN_PERMISSION)) ? this.confirmCompleteGoal.bind(this, this.state.model.data.id, false) : "")}>
-							<span className={this.state.model.data.closed == false ? "mdi mdi-lock-open-outline lockGoal-open":"mdi  mdi-lock lockGoal-close"}
-								title= {this.state.model.data.closed == false ? Messages.get("label.title.finishGoal") : ((this.context.roles.MANAGER || _.contains(this.context.permissions,
-	         					PermissionsTypes.MANAGE_PLAN_PERMISSION)) ? "Abrir Meta" : Messages.get("label.title.reopenGoal"))}>
-								<span id="menu-levels"> {this.state.model.data.closed == false ? Messages.getEditable("label.msg.finishGoal","fpdi-nav-label"):Messages.getEditable("label.msg.reopenGoal","fpdi-nav-label")} </span>
-							</span>
-						</a>
-					</li>) : ""}
+				</li>
 
 				{(this.context.roles.MANAGER || _.contains(this.context.permissions,
-         							PermissionsTypes.MANAGE_PLAN_PERMISSION)) && this.state.vizualization ? (
-			        <li>
-			        	<a onClick={this.exportLevelAttributes}>
-							<span className="mdi mdi-file-export cursorPointer" title={Messages.get("label.generateReport")}>
-								<span id="menu-levels"> {Messages.getEditable("label.generateReport","fpdi-nav-label")}</span>
-							</span>
-						</a>
-			        </li>) : ""}
-
-				{this.context.roles.SYSADMIN ? "" :
-			        (this.state.favoriteExistent ? (
-		         		<li>
-		         			<a onClick={this.removeFavoriteLevel}>
-								<span className="mdi mdi-star-outline cursorPointer" title={Messages.get("label.removeFromFavorites")}>
-									<span id="menu-levels"> {Messages.getEditable("label.removeFromFavorites","fpdi-nav-label")} </span>
+					PermissionsTypes.MANAGE_PLAN_PERMISSION)) && this.state.model.data.level.name == "Meta" ? (
+						<li>
+							<a onClick={this.state.model.data.closed == false ?
+								this.confirmCompleteGoal.bind(this, this.state.model.data.id, true) : ((this.context.roles.MANAGER || _.contains(this.context.permissions,
+									PermissionsTypes.MANAGE_PLAN_PERMISSION)) ? this.confirmCompleteGoal.bind(this, this.state.model.data.id, false) : "")}>
+								<span className={this.state.model.data.closed == false ? "mdi mdi-lock-open-outline lockGoal-open" : "mdi  mdi-lock lockGoal-close"}
+									title={this.state.model.data.closed == false ? Messages.get("label.title.finishGoal") : ((this.context.roles.MANAGER || _.contains(this.context.permissions,
+										PermissionsTypes.MANAGE_PLAN_PERMISSION)) ? "Abrir Meta" : Messages.get("label.title.reopenGoal"))}>
+									<span id="menu-levels"> {this.state.model.data.closed == false ? Messages.getEditable("label.msg.finishGoal", "fpdi-nav-label") : Messages.getEditable("label.msg.reopenGoal", "fpdi-nav-label")} </span>
 								</span>
 							</a>
-			         	</li>)
-			        	: (!this.state.favoriteTotal || this.state.favoriteTotal < 10 ?
-	         		 		<li>
-	         		 			<a onClick={this.saveFavoriteLevel}>
+						</li>) : ""}
+
+				{(this.context.roles.MANAGER || _.contains(this.context.permissions,
+					PermissionsTypes.MANAGE_PLAN_PERMISSION)) && this.state.vizualization ? (
+						<li>
+							<a onClick={this.exportLevelAttributes}>
+								<span className="mdi mdi-file-export cursorPointer" title={Messages.get("label.generateReport")}>
+									<span id="menu-levels"> {Messages.getEditable("label.generateReport", "fpdi-nav-label")}</span>
+								</span>
+							</a>
+						</li>) : ""}
+
+				{this.context.roles.SYSADMIN ? "" :
+					(this.state.favoriteExistent ? (
+						<li>
+							<a onClick={this.removeFavoriteLevel}>
+								<span className="mdi mdi-star-outline cursorPointer" title={Messages.get("label.removeFromFavorites")}>
+									<span id="menu-levels"> {Messages.getEditable("label.removeFromFavorites", "fpdi-nav-label")} </span>
+								</span>
+							</a>
+						</li>)
+						: (!this.state.favoriteTotal || this.state.favoriteTotal < 10 ?
+							<li>
+								<a onClick={this.saveFavoriteLevel}>
 									<span className="mdi mdi-star cursorPointer" title={Messages.get("label.addFavorites")}>
-										<span id="menu-levels">	{Messages.getEditable("label.addFavorites","fpdi-nav-label")} </span>
+										<span id="menu-levels">	{Messages.getEditable("label.addFavorites", "fpdi-nav-label")} </span>
 									</span>
 								</a>
-			         		</li>
-		         			:
-			         			<li>
-			         				<a type="submit">
-										<span className="mdi mdi-star disabledIcon" title={Messages.get("label.maxFavorites")}>
-											<span id="menu-levels">	{Messages.getEditable("label.notAddFavorites","fpdi-nav-label")} </span>
-										</span>
-									</a>
-					         	</li>
-		         		)
-		         	)
-			    }
+							</li>
+							:
+							<li>
+								<a type="submit">
+									<span className="mdi mdi-star disabledIcon" title={Messages.get("label.maxFavorites")}>
+										<span id="menu-levels">	{Messages.getEditable("label.notAddFavorites", "fpdi-nav-label")} </span>
+									</span>
+								</a>
+							</li>
+						)
+					)
+				}
 
-		         {this.state.undeletable ?
-		         	<li>
-		         		<a type="submit">
-							<span className="mdi mdi-delete disabledIcon marginLeft5" title={Messages.get("label.notDeleted") + " " +(!!this.state.model.data.closed ? ", meta concluída\n" : "")}>
-								<span id="menu-levels">	{Messages.getEditable("label.notDeleted","fpdi-nav-label")} </span>
+				{this.state.undeletable ?
+					<li>
+						<a type="submit">
+							<span className="mdi mdi-delete disabledIcon marginLeft5" title={Messages.get("label.notDeleted") + " " + (!!this.state.model.data.closed ? ", meta concluída\n" : "")}>
+								<span id="menu-levels">	{Messages.getEditable("label.notDeleted", "fpdi-nav-label")} </span>
 							</span>
 						</a>
-		         	</li>
-		         : ""}
+					</li>
+					: ""}
 
 			</ul>
 		);
 
-	 },
+	},
 
-	 renderDisableDelete() {
+	renderDisableDelete() {
 		return (
 			<ul id="level-menu" className="dropdown-menu">
 				<li>
 					<a>
 						<span className="mdi mdi-delete disabledIcon" title={Messages.get("label.unableDeleteArchivedPlan")}>
-							<span id="menu-levels">	{Messages.getEditable("label.unableDeleteArchivedPlan","fpdi-nav-label")} </span>
+							<span id="menu-levels">	{Messages.getEditable("label.unableDeleteArchivedPlan", "fpdi-nav-label")} </span>
 						</span>
 					</a>
-		         </li>
+				</li>
 			</ul>
 		);
 	},
@@ -924,10 +930,10 @@ export default React.createClass({
 				<li>
 					<a onClick={this.deleteLevelAttribute}>
 						<span className="mdi mdi-delete cursorPointer deleteIcon" title={Messages.get("label.delete")}>
-							<span id="menu-levels">	{Messages.getEditable("label.delete"),"fpdi-nav-label"} </span>
+							<span id="menu-levels">	{Messages.getEditable("label.delete", "fpdi-nav-label")} </span>
 						</span>
 					</a>
-		         </li>
+				</li>
 			</ul>
 		);
 
@@ -938,11 +944,11 @@ export default React.createClass({
 			<ul id="level-menu" className="dropdown-menu">
 				<li>
 					<a onClick={this.deleteLevelAttribute}>
-						<span className="mdi mdi-delete disabledIcon marginLeft5" title={Messages.get("label.notDeleted") +(!!this.state.model.data.closed ? ", meta concluída\n" : "")}>
-							<span id="menu-levels">	{Messages.getEditable("label.notDeleted","fpdi-nav-label")} </span>
+						<span className="mdi mdi-delete disabledIcon marginLeft5" title={Messages.get("label.notDeleted") + (!!this.state.model.data.closed ? ", meta concluída\n" : "")}>
+							<span id="menu-levels">	{Messages.getEditable("label.notDeleted", "fpdi-nav-label")} </span>
 						</span>
 					</a>
-		         </li>
+				</li>
 			</ul>
 		);
 	},
@@ -952,103 +958,103 @@ export default React.createClass({
 			return <LoadingGauge />;
 		}
 		var userResponsible;
-		for(var i = 0; i < this.state.fields.length; i++) {
-			if(this.state.fields[i].type == AttributeTypes.RESPONSIBLE_FIELD) {
+		for (var i = 0; i < this.state.fields.length; i++) {
+			if (this.state.fields[i].type == AttributeTypes.RESPONSIBLE_FIELD) {
 				if (this.state.fields[i].users) {
-					for(var j = 0; j < this.state.fields[i].users.length;j++) {
-						if(this.state.fields[i].users[j].id == this.state.fields[i].value) {
+					for (var j = 0; j < this.state.fields[i].users.length; j++) {
+						if (this.state.fields[i].users[j].id == this.state.fields[i].value) {
 							userResponsible = this.state.fields[i].users[j];
 						}
 					}
 				}
 			}
 		}
-		return(
-		<div>
-			<div className="row paddingLeft15">
-				{this.state.model ? this.renderBreadcrumb() : ""}
+		return (
+			<div>
+				<div className="row paddingLeft15">
+					{this.state.model ? this.renderBreadcrumb() : ""}
 
-				<div className={this.state.level == "Indicador" ? "col-lg-10 fpdi-indicatorTitle" : "col-lg-10 fpdi-attributesTitle"}>
-					<h1>
-						<div  className="fpdi-containerTitle" ref='levelInstanceTitle'>
-							{this.state.subTitle}
-						</div>
-						{this.state.vizualization ?
-						<div className="fpdi-containerIcons">
-							<span className="dropdown">
-								<a
-									className="dropdown-toggle"
-									data-toggle="dropdown"
-									aria-haspopup="true"
-									aria-expanded="true"
-									title={Messages.get("label.actions")}
-								>
-									<span className="sr-only">{Messages.getEditable("label.actions","fpdi-nav-label")}</span>
-									<span className="mdi mdi-chevron-down" />
-								</a>
-									{this.context.planMacro.attributes.archived ?  this.renderDisableEdit() :
-										(this.state.model.data.level.goal == false ? this.renderEditGoal(userResponsible) :
-				            				this.state.model.data.closed == false ? this.renderEditGoal(userResponsible) : this.renderClosedGoal())
-				            		}
-				            </span>
-						</div>:""}
-					</h1>
-					{this.state.level != "Indicador" ?
-						<div>
-							<span className="fpdi-level-type-value">
-								{this.state.level}
-							</span>
-						</div>
-					:
-						this.state.aggregate ?
+					<div className={this.state.level == "Indicador" ? "col-lg-10 fpdi-indicatorTitle" : "col-lg-10 fpdi-attributesTitle"}>
+						<h1>
+							<div className="fpdi-containerTitle" ref='levelInstanceTitle'>
+								{this.state.subTitle}
+							</div>
+							{this.state.vizualization ?
+								<div className="fpdi-containerIcons">
+									<span className="dropdown">
+										<a
+											className="dropdown-toggle"
+											data-toggle="dropdown"
+											aria-haspopup="true"
+											aria-expanded="true"
+											title={Messages.get("label.actions")}
+										>
+											<span className="sr-only">{Messages.getEditable("label.actions", "fpdi-nav-label")}</span>
+											<span className="mdi mdi-chevron-down" />
+										</a>
+										{this.context.planMacro.attributes.archived ? this.renderDisableEdit() :
+											(this.state.model.data.level.goal == false ? this.renderEditGoal(userResponsible) :
+												this.state.model.data.closed == false ? this.renderEditGoal(userResponsible) : this.renderClosedGoal())
+										}
+									</span>
+								</div> : ""}
+						</h1>
+						{this.state.level != "Indicador" ?
 							<div>
 								<span className="fpdi-level-type-value">
-									{Messages.getEditable("label.indicatorAggregate","fpdi-nav-label")}
+									{this.state.level}
 								</span>
 							</div>
-						:
-							<div>
-								<span className="fpdi-level-type-value">
-									{Messages.getEditable("label.indicatorSimple","fpdi-nav-label")}
-								</span>
-							</div>
-					}
-				</div>
-				<div className = "fpdi-levelValueBox">
-					<div className = {this.state.levelValue == null || this.state.levelValue < 1000 ? "col-lg-1 box-proceeds" : "col-lg-1"} >
-						<span id = "title-proceeds"> {Messages.getEditable("label.profit","fpdi-nav-label")} </span>
-						<span id = "value-proceeds"> {this.state.levelValue == null ? 0 + "%" : this.state.levelValue.toFixed(2)+"%"} </span>
+							:
+							this.state.aggregate ?
+								<div>
+									<span className="fpdi-level-type-value">
+										{Messages.getEditable("label.indicatorAggregate", "fpdi-nav-label")}
+									</span>
+								</div>
+								:
+								<div>
+									<span className="fpdi-level-type-value">
+										{Messages.getEditable("label.indicatorSimple", "fpdi-nav-label")}
+									</span>
+								</div>
+						}
+					</div>
+					<div className="fpdi-levelValueBox">
+						<div className={this.state.levelValue == null || this.state.levelValue < 1000 ? "col-lg-1 box-proceeds" : "col-lg-1"} >
+							<span id="title-proceeds"> {Messages.getEditable("label.profit", "fpdi-nav-label")} </span>
+							<span id="value-proceeds"> {this.state.levelValue == null ? 0 + "%" : this.state.levelValue.toFixed(2) + "%"} </span>
+						</div>
 					</div>
 				</div>
-			</div>
 
-			<VerticalForm
-			    ref="levelForm"
-			    vizualization={this.state.vizualization}
-				isPlan={true}
-			    onCancel={this.onCancel}
-				onSubmit={this.onSubmit}
-				fields={this.state.fields}
-				store={StructureStore}
-				onChange={this.onChange}
-				submitLabel="Salvar"
-				dateBegin={dateBegin}
-				dateEnd={dateEnd}
-				userResponsible={userResponsible}
-				levelInstanceId={this.state.model.data.id}
-				levelInstanceIdActionPlan = {this.props.params.levelInstanceId}/>
+				<VerticalForm
+					ref="levelForm"
+					vizualization={this.state.vizualization}
+					isPlan={true}
+					onCancel={this.onCancel}
+					onSubmit={this.onSubmit}
+					fields={this.state.fields}
+					store={StructureStore}
+					onChange={this.onChange}
+					submitLabel="Salvar"
+					dateBegin={dateBegin}
+					dateEnd={dateEnd}
+					userResponsible={userResponsible}
+					levelInstanceId={this.state.model.data.id}
+					levelInstanceIdActionPlan={this.props.params.levelInstanceId} />
 
 
-			{this.state.showPolarityAlert ? <i><h5 className="fpdi-indicator-weigth-total">{Messages.getEditable("label.msg.polarityChanged")}</h5></i> : ""}
-			{this.state.model.data.level.indicator ? <GoalPerformance  indicator = {this.state.model.data}/> : ""}
+				{this.state.showPolarityAlert ? <i><h5 className="fpdi-indicator-weigth-total">{Messages.getEditable("label.msg.polarityChanged")}</h5></i> : ""}
+				{this.state.model.data.level.indicator ? <GoalPerformance indicator={this.state.model.data} /> : ""}
 
-			<LevelSons
-				dateBegin={dateBegin}
-				dateEnd={dateEnd}
-				parentId={this.state.model.data.id}
-				isParent={this.isParent}
-				users={this.state.users}
-				enabled={!this.state.aggregate && this.state.vizualization}/>
-	    </div>);
+				<LevelSons
+					dateBegin={dateBegin}
+					dateEnd={dateEnd}
+					parentId={this.state.model.data.id}
+					isParent={this.isParent}
+					users={this.state.users}
+					enabled={!this.state.aggregate && this.state.vizualization} />
+			</div>);
 	}
 });
