@@ -2,7 +2,7 @@ import _ from "underscore";
 import React from "react";
 import MaskedInput from 'react-maskedinput';
 import RichText from 'forpdi/jsx/vendor/FPDIRichText.jsx';
-import {Link} from 'react-router';
+import { Link } from 'react-router';
 import DatePicker from 'react-datepicker';
 import moment from 'moment';
 import 'react-datepicker/dist/react-datepicker.css';
@@ -20,6 +20,8 @@ import SelectStructure from "forpdi/jsx/planning/view/field/SelectStructureField
 import PlanMacroStore from "forpdi/jsx/planning/store/PlanMacro.jsx";
 import Messages from "forpdi/jsx/core/util/Messages.jsx";
 
+import Select from 'react-select';
+
 export default React.createClass({
 	contextTypes: {
 		toastr: React.PropTypes.object.isRequired,
@@ -32,7 +34,7 @@ export default React.createClass({
 		return {
 			fieldDef: {
 				type: "text",
-				name: "input-"+Date.now(),
+				name: "input-" + Date.now(),
 				label: '',
 				placeholder: '',
 				value: null,
@@ -41,32 +43,38 @@ export default React.createClass({
 				undeletable: false,
 				alterable: false,
 				editing: false
-			}
+			},
+			selectedOption: null,
+			optionsField:[],
 		};
 	},
 	getInitialState() {
 		return {
-			fieldId: "field-"+this.props.fieldDef.name.replace(/\./g, "-"),
+			fieldId: "field-" + this.props.fieldDef.name.replace(/\./g, "-"),
 			strategicObjectivesPlansParam: -1
 		};
 	},
 	getValue() {
 		var el = this.refs[this.state.fieldId];
 
-		if (el == undefined)
+		if (!el)
 			return el;
-		if (el.type == AttributeTypes.DATE_FIELD || 
-				el.type ==  AttributeTypes.DATE_TIME_FIELD)
+		if (!el.type) {
+			// para o campo de text area o type esta undefined
+			return el.state.value;
+		}
+		if (el.type == AttributeTypes.DATE_FIELD ||
+			el.type == AttributeTypes.DATE_TIME_FIELD)
 			return el.valueAsDate;
-		if (el.type == AttributeTypes.CURRENCY_FIELD || 
-				el.type == AttributeTypes.NUMBER_FIELD ||
-				el.type == AttributeTypes.PERCENTAGE_FIELD ||
-				el.type == AttributeTypes.TOTAL_FIELD)
+		if (el.type == AttributeTypes.CURRENCY_FIELD ||
+			el.type == AttributeTypes.NUMBER_FIELD ||
+			el.type == AttributeTypes.PERCENTAGE_FIELD ||
+			el.type == AttributeTypes.TOTAL_FIELD)
 			return el.valueAsNumber;
 		if (this.props.fieldDef.type == AttributeTypes.SELECT_PLAN_FIELD) {
 			if (this.props.fieldDef.selectPlans.length > 0) {
-				for (var i=0; i<this.props.fieldDef.selectPlans.length; i++) {
-					if(this.props.fieldDef.selectPlans[i].name == el.value) {
+				for (var i = 0; i < this.props.fieldDef.selectPlans.length; i++) {
+					if (this.props.fieldDef.selectPlans[i].name == el.value) {
 						return this.props.fieldDef.selectPlans[i].id.toString();
 					}
 				}
@@ -82,31 +90,31 @@ export default React.createClass({
 	getInputNode() {
 		return this.refs[this.state.fieldId];
 	},
-	onKeyUp(evt){
+	onKeyUp(evt) {
 		this.maxLengthMask();
 		var key = evt.which;
-		if(key == 13) {			
+		if (key == 13) {
 			evt.preventDefault();
 			return;
 		}
-		if(this.refs[this.state.fieldId].value.length+1 > this.refs[this.state.fieldId].maxLength) {
-			this.refs[this.state.fieldId].value = this.refs[this.state.fieldId].value.substr(0, this.refs[this.state.fieldId].maxLength-1);
+		if (this.refs[this.state.fieldId].value.length + 1 > this.refs[this.state.fieldId].maxLength) {
+			this.refs[this.state.fieldId].value = this.refs[this.state.fieldId].value.substr(0, this.refs[this.state.fieldId].maxLength - 1);
 		}
 	},
-	maxLengthMask(){
-		if(this.refs[this.state.fieldId].value.length >= this.refs[this.state.fieldId].maxLength){
-			this.context.toastr.addAlertError( Messages.get("label.error.limit") + " " +this.refs[this.state.fieldId].maxLength+" " + Messages.get("label.error.limitCaracteres"));
+	maxLengthMask() {
+		if (this.refs[this.state.fieldId].value.length >= this.refs[this.state.fieldId].maxLength) {
+			this.context.toastr.addAlertError(Messages.get("label.error.limit") + " " + this.refs[this.state.fieldId].maxLength + " " + Messages.get("label.error.limitCaracteres"));
 		}
 	},
-	onlyNumber(evt){
- 		var key = evt.which;
- 		if(key == 13|| key != 44 && (key < 48 || key > 57)) {
- 			evt.preventDefault();
- 			return;
- 		}
- 	},
-	
-	onlyNumberPaste(evt){
+	onlyNumber(evt) {
+		var key = evt.which;
+		if (key == 13 || key != 44 && (key < 48 || key > 57)) {
+			evt.preventDefault();
+			return;
+		}
+	},
+
+	onlyNumberPaste(evt) {
 		var value = evt.clipboardData.getData('Text');
 		if (!(!isNaN(parseFloat(value)) && isFinite(value)) || parseFloat(value) < 0) {
 			evt.preventDefault();
@@ -114,96 +122,100 @@ export default React.createClass({
 		}
 	},
 
-	componentDidMount() {		
-		if (this.props.fieldDef.type == AttributeTypes.DATE_FIELD || 
-				this.props.fieldDef.type == AttributeTypes.DATE_TIME_FIELD) {
+	componentDidMount() {
+		if (this.props.fieldDef.type == AttributeTypes.DATE_FIELD ||
+			this.props.fieldDef.type == AttributeTypes.DATE_TIME_FIELD) {
 			$(this.getInputNode()).daterangepicker({
 				autoApply: true,
 				autoUpdateInput: true,
 				locale: {
-		            format: 'DD/MM/YYYY'
-		        },
-		        opens: 'right',
-		        drops: 'down',
-		        showDropdowns: true,
-		        singleDatePicker: true
+					format: 'DD/MM/YYYY'
+				},
+				opens: 'right',
+				drops: 'down',
+				showDropdowns: true,
+				singleDatePicker: true
 			});
 		}
 
-		PlanMacroStore.on("getmainmenustate", (data) => {	
-            this.setState({
-                menuHidden: data
-            });
-        }, this);	
+		PlanMacroStore.on("getmainmenustate", (data) => {
+			this.setState({
+				menuHidden: data
+			});
+		}, this);
 	},
 
-	onStrategicObjectivesSelectPlanChange(){
-      this.setState({
-        strategicObjectivesPlansParam: this.refs.strategicObjectivesSelectPlan.value
-      })
-      
-    },
+	onStrategicObjectivesSelectPlanChange() {
+		this.setState({
+			strategicObjectivesPlansParam: this.refs.strategicObjectivesSelectPlan.value
+		})
 
-	delete(){
+	},
+
+	delete() {
 		this.props.deleteFunc(this.props.id);
 	},
 
-	edit(){
+	edit() {
 		this.setState({
 			editing: true
 		});
 	},
 
-	cancelEditing(){
+	cancelEditing() {
 		this.setState({
 			editing: false
 		});
 	},
 
-	confirmEdit(){
-		if(this.refs['edit-input'].value.trim() != ""){
+	confirmEdit() {
+		if (this.refs['edit-input'].value.trim() != "") {
 			this.props.editFunc(this.refs['edit-input'].value, this.props.index);
 			this.cancelEditing();
-		}else{
+		} else {
 			this.context.toastr.addAlertError(Messages.get("label.completedField"));
 		}
 	},
 
-	renderEditing(){
-		return(
-			<div className="edit-section-attribute"> 
-				<input defaultValue={this.props.fieldDef.label} className="edit-section-attribute-input" maxLength="255" ref="edit-input"/>
+	renderEditing() {
+		return (
+			<div className="edit-section-attribute">
+				<input defaultValue={this.props.fieldDef.label} className="edit-section-attribute-input" maxLength="255" ref="edit-input" />
 				<div className='displayFlex'>
-                   	<span className='mdi mdi-check accepted-budget' onClick={this.confirmEdit} title={Messages.get("label.submitLabel")}></span>
-                  	<span className='mdi mdi-close reject-budget' onClick={this.cancelEditing} title={Messages.get("label.cancel")}></span>
-               	</div>
+					<span className='mdi mdi-check accepted-budget' onClick={this.confirmEdit} title={Messages.get("label.submitLabel")}></span>
+					<span className='mdi mdi-close reject-budget' onClick={this.cancelEditing} title={Messages.get("label.cancel")}></span>
+				</div>
 			</div>
 		);
 	},
 
-	renderLabel(param){
-		return(
+	renderLabel(param) {
+		return (
 			<div>
 				<b className="budget-title">{this.props.fieldDef.label}</b>
 				{param}
-				{(this.context.roles.MANAGER || _.contains(this.context.permissions, 
-         				PermissionsTypes.MANAGE_DOCUMENT_PERMISSION)) && !this.context.planMacro.get("archived")?
+				{(this.context.roles.MANAGER || _.contains(this.context.permissions,
+					PermissionsTypes.MANAGE_DOCUMENT_PERMISSION)) && !this.context.planMacro.get("archived") ?
 					(!!this.props.undeletable ? <span type="submit" className="mdi mdi-delete attribute-input-edit inner"
-				 		title={Messages.get("label.title.deleteField")} onClick={this.delete}/> : "")
-				: ""}
-				{(this.context.roles.MANAGER || _.contains(this.context.permissions, 
-         				PermissionsTypes.MANAGE_DOCUMENT_PERMISSION)) && !this.context.planMacro.get("archived")?
-					(!!this.props.alterable ? <span className="mdi mdi-pencil attribute-input-edit inner" 
-							title={Messages.get("label.title.changeField")} onClick={this.edit}/> : "")
-				:""}
+						title={Messages.get("label.title.deleteField")} onClick={this.delete} /> : "")
+					: ""}
+				{(this.context.roles.MANAGER || _.contains(this.context.permissions,
+					PermissionsTypes.MANAGE_DOCUMENT_PERMISSION)) && !this.context.planMacro.get("archived") ?
+					(!!this.props.alterable ? <span className="mdi mdi-pencil attribute-input-edit inner"
+						title={Messages.get("label.title.changeField")} onClick={this.edit} /> : "")
+					: ""}
 			</div>
 		);
 	},
 
+	handleChange(selectedOption){
+		this.setState({ selectedOption });
+	  },
+
 	render() {
-		var fieldEl;		
+		var fieldEl;
 		if (this.props.vizualization) {
-			if (this.props.fieldDef.type == AttributeTypes.INDICATOR_TYPE){	
+			if (this.props.fieldDef.type == AttributeTypes.INDICATOR_TYPE) {
 				fieldEl = (
 					<IndicatorType fieldDef={this.props.fieldDef} vizualization={this.props.vizualization} />
 				);
@@ -219,27 +231,33 @@ export default React.createClass({
 				fieldEl = (
 					<SelectPlan fieldDef={this.props.fieldDef} vizualization={this.props.vizualization} />
 				);
-			} else if ((this.props.fieldDef.type == AttributeTypes.NUMBER_FIELD || 
+			} else if ((this.props.fieldDef.type == AttributeTypes.NUMBER_FIELD ||
 				this.props.fieldDef.type == AttributeTypes.CURRENCY_FIELD ||
-				this.props.fieldDef.type == AttributeTypes.PERCENTAGE_FIELD) && this.props.fieldDef.formattedValue){
+				this.props.fieldDef.type == AttributeTypes.PERCENTAGE_FIELD) && this.props.fieldDef.formattedValue) {
 				fieldEl = (
-					<span className="pdi-normal-text">{this.props.fieldDef.formattedValue.trim().substr(0,4) == "null" ? "" : this.props.fieldDef.formattedValue}</span>
+					<span className="pdi-normal-text">{this.props.fieldDef.formattedValue.trim().substr(0, 4) == "null" ? "" : this.props.fieldDef.formattedValue}</span>
 				);
 			} else if (this.props.fieldDef.type == AttributeTypes.STRATEGIC_OBJECTIVE_FIELD) {
 				fieldEl = (
 					<StrategicObjective fieldId={this.state.fieldId} fieldDef={this.props.fieldDef} strategicObjectivesPlansParam={this.state.strategicObjectivesPlansParam} />
-				);		
+				);
 			} else {
- 				fieldEl = (
- 					<div><span className="pdi-normal-text" dangerouslySetInnerHTML={{__html: this.props.fieldDef.value}}/></div>
- 				);
- 			}
+				if (this.props.fieldDef.description != null) {
+					fieldEl = (
+						<div><span className="pdi-normal-text" dangerouslySetInnerHTML={{ __html: this.props.fieldDef.description }} /></div>
+					);
+				} else {
+					fieldEl = (
+						<div><span className="pdi-normal-text" dangerouslySetInnerHTML={{ __html: this.props.fieldDef.value }} /></div>
+					);
+				}
+			}
 		} else if (this.props.fieldDef.type == AttributeTypes.STRATEGIC_OBJECTIVE_FIELD) {
 			fieldEl = (
 				<StrategicObjective fieldId={this.state.fieldId} fieldDef={this.props.fieldDef} strategicObjectivesPlansParam={this.state.strategicObjectivesPlansParam} />
 			);
- 		} else if (this.props.fieldDef.type == AttributeTypes.TEXT_AREA_FIELD) {
-			if(this.props.isDocument){
+		} else if (this.props.fieldDef.type == AttributeTypes.TEXT_AREA_FIELD) {
+			if (this.props.isDocument) {
 				fieldEl = (
 					<div>
 						<RichText
@@ -256,10 +274,10 @@ export default React.createClass({
 							onPaste={this.onKeyUp}
 						/>
 						<div className="textAreaMaxLenght documentText">
-							<span>{Messages.getEditable("label.maxTenThousandCaracteres","fpdi-nav-label")}</span>
+							<span>{Messages.getEditable("label.maxTenThousandCaracteres", "fpdi-nav-label")}</span>
 						</div>
 					</div>
-				);			
+				);
 			} else {
 				if (this.props.fieldDef.disabled) {
 					fieldEl = (
@@ -281,7 +299,7 @@ export default React.createClass({
 								title={Messages.get("label.haveNoPermissionToEdit")}
 							/>
 							<div className="textAreaMaxLenght">
-								<span>{Messages.getEditable("label.fourThousandCaracteres","fpdi-nav-label")}</span>
+								<span>{Messages.getEditable("label.fourThousandCaracteres", "fpdi-nav-label")}</span>
 							</div>
 						</div>
 					);
@@ -303,15 +321,43 @@ export default React.createClass({
 								onPaste={this.onKeyUp}
 							/>
 							<div className="textAreaMaxLenght">
-								<span>{Messages.getEditable("label.fourThousandCaracteres","fpdi-nav-label")}</span>
+								<span>{Messages.getEditable("label.fourThousandCaracteres", "fpdi-nav-label")}</span>
 							</div>
 						</div>
 					);
 				}
 			}
 		} else if (this.props.fieldDef.type == AttributeTypes.SELECT_FIELD) {
-			fieldEl = (
-				<select
+
+			if (this.props.fieldDef.search) {
+				if(this.state.selectedOption == null){
+					var selected=[]
+					var options=[]
+
+					for(var i=0; i<this.props.fieldDef.optionsField.length;i++){
+						options.push({label:this.props.fieldDef.optionsField[i].label, value:this.props.fieldDef.optionsField[i].id})
+					}
+					this.state.optionsField=options
+
+					if(this.props.fieldDef.value != null){
+						this.state.selectedOption={label:this.props.fieldDef.value, value:this.props.fieldDef.id}
+					}
+				}
+
+				fieldEl = (<Select
+					value={this.state.selectedOption}
+					onChange={this.handleChange}
+					name={this.props.fieldDef.name}
+					id={this.state.fieldId}
+					ref={this.state.fieldId}
+					type={this.props.fieldDef.type}
+					options={this.state.optionsField}
+					placeholder={this.props.fieldDef.placeholder}
+					isSearchable="true"
+				/>)
+
+			} else {
+				fieldEl = (<select
 					className="form-control"
 					placeholder={this.props.fieldDef.placeholder}
 					name={this.props.fieldDef.name}
@@ -320,14 +366,18 @@ export default React.createClass({
 					type={this.props.fieldDef.type}
 					onChange={this.props.fieldDef.onChange || _.noop}
 					defaultValue={this.props.fieldDef.value}
-					>
-						{this.props.fieldDef.optionsField ? this.props.fieldDef.optionsField.map((opt,idx) => {
-							return (<option key={'field-opt-'+this.state.fieldId+"-"+idx} defaultValue={opt.get ? opt.label : ""}
-								data-placement="right" title={opt.label}>
-									{opt.label}</option>);
-						}):''}
-				</select>
-			);	
+				>
+				{this.props.fieldDef.value == null ?
+					<option value="" disabled selected>{this.props.fieldDef.placeholder}</option >
+					: <option value="" disabled>{this.props.fieldDef.placeholder}</option >
+				}
+				{this.props.fieldDef.optionsField ? this.props.fieldDef.optionsField.map((opt, idx) => {
+					return (<option key={'field-opt-' + this.state.fieldId + "-" + idx} defaultValue={opt.label ? opt.label : null}
+						data-placement="right" title={opt.label}>
+						{opt.label}</option>);
+				}) : ''}
+				</select>);
+			}
 		} else if (this.props.fieldDef.type == AttributeTypes.RESPONSIBLE_FIELD) {
 			if (!this.props.fieldDef.users || this.props.fieldDef.users.length <= 0) {
 				fieldEl = (
@@ -340,7 +390,7 @@ export default React.createClass({
 						type={this.props.fieldDef.type}
 						disabled
 						title={Messages.get("label.needRegisterUsersToProceed")}
-						>
+					>
 					</input>
 				);
 			} else if (this.props.fieldDef.disabled) {
@@ -356,15 +406,15 @@ export default React.createClass({
 						defaultValue={this.props.fieldDef.value}
 						disabled
 						title={Messages.get("label.haveNoPermissionToEdit")}
-						>
-							{this.props.fieldDef.users ? this.props.fieldDef.users.map((user,idx) => {
-								return (<option key={'field-opt-'+this.state.fieldId+"-"+idx} value={user.id}
-									data-placement="right" title={user.name}>
-										{user.name}</option>);
-							}):''}
+					>
+						{this.props.fieldDef.users ? this.props.fieldDef.users.map((user, idx) => {
+							return (<option key={'field-opt-' + this.state.fieldId + "-" + idx} value={user.id}
+								data-placement="right" title={user.name}>
+								{user.name}</option>);
+						}) : ''}
 					</select>
 				);
-			} else {				
+			} else {
 				fieldEl = (
 					<select
 						className="form-control"
@@ -375,12 +425,12 @@ export default React.createClass({
 						type={this.props.fieldDef.type}
 						onChange={this.props.fieldDef.onChange || _.noop}
 						defaultValue={this.props.fieldDef.value}
-						>
-							{this.props.fieldDef.users ? this.props.fieldDef.users.map((user,idx) => {
-								return (<option key={'field-opt-'+this.state.fieldId+"-"+idx} value={user.id}
-									data-placement="right" title={user.name}>
-										{user.name}</option>);
-							}):''}
+					>
+						{this.props.fieldDef.users ? this.props.fieldDef.users.map((user, idx) => {
+							return (<option key={'field-opt-' + this.state.fieldId + "-" + idx} value={user.id}
+								data-placement="right" title={user.name}>
+								{user.name}</option>);
+						}) : ''}
 					</select>
 				);
 			}
@@ -396,13 +446,13 @@ export default React.createClass({
 						type={this.props.fieldDef.type}
 						onChange={this.props.fieldDef.onChange || _.noop}
 						defaultValue={this.props.fieldDef.value}
-						>
-							{this.props.fieldDef.selectPlans.map((opt,idx) => {
-								return (<option key={'field-opt-'+this.state.fieldId+"-"+idx} value={opt.id}
-									data-placement="right" title={opt.name}>
-									 {(opt.name.length>92)?(string(opt.name).trim().substr(0, 89).concat("...").toString()):(opt.name)}
-									</option>);
-							})}
+					>
+						{this.props.fieldDef.selectPlans.map((opt, idx) => {
+							return (<option key={'field-opt-' + this.state.fieldId + "-" + idx} value={opt.id}
+								data-placement="right" title={opt.name}>
+								{(opt.name.length > 92) ? (string(opt.name).trim().substr(0, 89).concat("...").toString()) : (opt.name)}
+							</option>);
+						})}
 					</select>
 				);
 			} else {
@@ -414,42 +464,42 @@ export default React.createClass({
 						id={this.state.fieldId}
 						ref={this.state.fieldId}
 						type={this.props.fieldDef.type}
-						>
+					>
 						<option key='field-opt-' data-placement="right" title={Messages.get("label.NoGoalPlanRegistered")}>
-							{Messages.getEditable("label.NoGoalPlanRegistered","fpdi-nav-label")}
+							{Messages.getEditable("label.NoGoalPlanRegistered", "fpdi-nav-label")}
 						</option>
 					</select>
 				);
 			}
-		} else if (this.props.fieldDef.type == AttributeTypes.RADIO || this.props.fieldDef.type == AttributeTypes.INDICATOR_TYPE) {			
+		} else if (this.props.fieldDef.type == AttributeTypes.RADIO || this.props.fieldDef.type == AttributeTypes.INDICATOR_TYPE) {
 			fieldEl = (
 				<div>
 					<div className="row">
-						{this.props.fieldDef.options.map((option, idx) => {		
-							return (															
-								<div className="fpdi-indicator-type-ctn col-sm-2" key={this.props.fieldDef.name+"-option-"+idx}>
+						{this.props.fieldDef.options.map((option, idx) => {
+							return (
+								<div className="fpdi-indicator-type-ctn col-sm-2" key={this.props.fieldDef.name + "-option-" + idx}>
 									<label onClick={this.props.fieldDef.onClick}>
 										<input
 											className="col-sm-6"
 											type="radio"
-											name={this.props.fieldDef.name+idx}
+											name={this.props.fieldDef.name + idx}
 											defaultValue={option[this.props.fieldDef.valueField]}
-											defaultChecked={option[this.props.fieldDef.valueField]}	
-											ref={this.props.fieldDef.name+"-option-"+idx}
+											defaultChecked={option[this.props.fieldDef.valueField]}
+											ref={this.props.fieldDef.name + "-option-" + idx}
 											onChange={this.props.fieldDef.onChange || _.noop}
 											onKeyPress={this.onKeyUp}
 											onPaste={this.onKeyUp}
-											onClick={this.props.fieldDef.onClick}/>
-										<label className="fpdi-indicator-type-label col-sm-6" id={"label-"+idx}>{option[this.props.fieldDef.displayField]}</label>
+											onClick={this.props.fieldDef.onClick} />
+										<label className="fpdi-indicator-type-label col-sm-6" id={"label-" + idx}>{option[this.props.fieldDef.displayField]}</label>
 									</label>
-								</div>							
+								</div>
 							);
-						})}						
+						})}
 					</div>
 					{this.props.fieldDef.extraRender ? this.props.fieldDef.extraRender() : ""}
 				</div>
 			);
-		} else if (this.props.fieldDef.type == AttributeTypes.SELECT_STRUCTURE) {		
+		} else if (this.props.fieldDef.type == AttributeTypes.SELECT_STRUCTURE) {
 			fieldEl = (
 				<select
 					className="form-control"
@@ -459,12 +509,12 @@ export default React.createClass({
 					ref={this.props.fieldDef.name}
 					onChange={this.props.fieldDef.onChange}
 					defaultValue={this.props.fieldDef.value}
-					>
-					{this.props.fieldDef.options.map((opt,idx) => {
-						return (<option key={'field-opt-'+opt.id+"-"+idx} value={opt.id}
+				>
+					{this.props.fieldDef.options.map((opt, idx) => {
+						return (<option key={'field-opt-' + opt.id + "-" + idx} value={opt.id}
 							data-placement="right" title={opt.name}>
-								 {(opt.get("name").length>20)?(string(opt.get("name")).trim().substr(0, 15).concat("...").toString()):(opt.get("name"))}
-								</option>);
+							{(opt.get("name").length > 20) ? (string(opt.get("name")).trim().substr(0, 15).concat("...").toString()) : (opt.get("name"))}
+						</option>);
 					})}
 				</select>
 			);
@@ -476,15 +526,15 @@ export default React.createClass({
 					name={this.props.fieldDef.name}
 					defaultValue={this.props.fieldDef.value}
 					ref={this.state.fieldId}
-					placeholderText="DD/MM/AAAA" 
+					placeholderText="DD/MM/AAAA"
 					dateFormat="DD/MM/YYYY"
 					id={this.state.fieldId}
-					selected={(this.props.fieldDef.value)?(moment(this.props.fieldDef.value, "DD/MM/YYYY")):(null)}
+					selected={(this.props.fieldDef.value) ? (moment(this.props.fieldDef.value, "DD/MM/YYYY")) : (null)}
 					onChange={this.props.fieldDef.onChange || _.noop}
 					onKeyPress={this.onKeyUp}
 					onPaste={this.onKeyUp}
 					maxLength='255'
-					showYearDropdown  
+					showYearDropdown
 					disabled
 					title={Messages.get("label.haveNoPermissionToEdit")}
 				/></div>);
@@ -495,10 +545,10 @@ export default React.createClass({
 					name={this.props.fieldDef.name}
 					defaultValue={this.props.fieldDef.value}
 					ref={this.state.fieldId}
-					placeholderText="DD/MM/AAAA" 
+					placeholderText="DD/MM/AAAA"
 					dateFormat="DD/MM/YYYY"
 					id={this.state.fieldId}
-					selected={(this.props.fieldDef.value)?(moment(this.props.fieldDef.value, "DD/MM/YYYY")):(null)}
+					selected={(this.props.fieldDef.value) ? (moment(this.props.fieldDef.value, "DD/MM/YYYY")) : (null)}
 					onChange={this.props.fieldDef.onChange || _.noop}
 					onKeyPress={this.onKeyUp}
 					onPaste={this.onKeyUp}
@@ -506,57 +556,57 @@ export default React.createClass({
 					showYearDropdown
 				/></div>);
 			}
-		} else if (this.props.fieldDef.type == AttributeTypes.NUMBER_FIELD || 
+		} else if (this.props.fieldDef.type == AttributeTypes.NUMBER_FIELD ||
 			this.props.fieldDef.type == AttributeTypes.PERCENTAGE_FIELD ||
-			this.props.fieldDef.type == AttributeTypes.CURRENCY_FIELD) {		
-			if (this.props.fieldDef.disabled) {	
-	 		 	fieldEl = (<input
-	 				onKeyPress={this.onlyNumber}
-					onPaste={this.onlyNumberPaste}	
-					className="budget-field-table"
-	 			//	type='number'
-	 				step='any'
-					min="0" //caso seja necessário campo numérico negativo, remover aqui!
-	 				name={this.props.fieldDef.name}
-	 				defaultValue={this.props.fieldDef.editModeValue}
-	 				id={this.state.fieldId}
-	 				ref={this.state.fieldId}
-	 				placeholder={this.props.fieldDef.placeholder}
-	 				onChange={this.props.fieldDef.onChange || _.noop}
-	 				disabled
-	 				title={Messages.get("label.haveNoPermissionToEdit")}
-	 			/>);
-	 		 } else {
-	 		 	fieldEl = (<input
-	 				onKeyPress={this.onlyNumber}
+			this.props.fieldDef.type == AttributeTypes.CURRENCY_FIELD) {
+			if (this.props.fieldDef.disabled) {
+				fieldEl = (<input
+					onKeyPress={this.onlyNumber}
 					onPaste={this.onlyNumberPaste}
-	 			//	className="budget-field-table"
-	 				className="form-control"
-	 				step='any'
-	 			//	type='number'
+					className="budget-field-table"
+					//	type='number'
+					step='any'
 					min="0" //caso seja necessário campo numérico negativo, remover aqui!
-	 				name={this.props.fieldDef.name}
-	 				defaultValue={this.props.fieldDef.editModeValue}
-	 				id={this.state.fieldId}
-	 				ref={this.state.fieldId}
-	 				placeholder={this.props.fieldDef.placeholder}
-	 				onChange={this.props.fieldDef.onChange || _.noop}
-	 			/>);
-	 		 }
-		} else if (this.props.fieldDef.type == AttributeTypes.TOTAL_FIELD){
+					name={this.props.fieldDef.name}
+					defaultValue={this.props.fieldDef.editModeValue}
+					id={this.state.fieldId}
+					ref={this.state.fieldId}
+					placeholder={this.props.fieldDef.placeholder}
+					onChange={this.props.fieldDef.onChange || _.noop}
+					disabled
+					title={Messages.get("label.haveNoPermissionToEdit")}
+				/>);
+			} else {
+				fieldEl = (<input
+					onKeyPress={this.onlyNumber}
+					onPaste={this.onlyNumberPaste}
+					//	className="budget-field-table"
+					className="form-control"
+					step='any'
+					//	type='number'
+					min="0" //caso seja necessário campo numérico negativo, remover aqui!
+					name={this.props.fieldDef.name}
+					defaultValue={this.props.fieldDef.editModeValue}
+					id={this.state.fieldId}
+					ref={this.state.fieldId}
+					placeholder={this.props.fieldDef.placeholder}
+					onChange={this.props.fieldDef.onChange || _.noop}
+				/>);
+			}
+		} else if (this.props.fieldDef.type == AttributeTypes.TOTAL_FIELD) {
 			fieldEl = (<input
- 				onKeyPress={this.onlyNumber}
+				onKeyPress={this.onlyNumber}
 				onPaste={this.onlyNumberPaste}
- 				className="form-control"
- 				type='number'
- 				name={this.props.fieldDef.name}
- 				value={this.props.fieldDef.value || 0}
- 				id={this.state.fieldId}
- 				ref={this.state.fieldId}
- 				placeholder={this.props.fieldDef.placeholder}
- 				onChange={this.props.fieldDef.onChange || _.noop} 				
+				className="form-control"
+				type='number'
+				name={this.props.fieldDef.name}
+				value={this.props.fieldDef.value || 0}
+				id={this.state.fieldId}
+				ref={this.state.fieldId}
+				placeholder={this.props.fieldDef.placeholder}
+				onChange={this.props.fieldDef.onChange || _.noop}
 				disabled={true}
- 			/>);
+			/>);
 		} else {
 			fieldEl = (<input
 				//className="budget-field-table"
@@ -577,7 +627,7 @@ export default React.createClass({
 				title={this.props.fieldDef.disabled ? Messages.get("label.haveNoPermissionToEdit") : ""}
 			/>);
 		}
-			
+
 		var strategicObjectivesPlans = "";
 		if (this.props.fieldDef.type == AttributeTypes.STRATEGIC_OBJECTIVE_FIELD) {
 			if (this.props.fieldDef.selectPlans.length > 0) {
@@ -591,15 +641,15 @@ export default React.createClass({
 						type={this.props.fieldDef.type}
 						onChange={this.onStrategicObjectivesSelectPlanChange}
 						defaultValue={this.props.fieldDef.value}
-						>
+					>
 						<option value={-1} data-placement="right" title={Messages.get("label.allGoalPlans")}>{Messages.get("label.allGoalPlans")} </option>
-							{this.props.fieldDef.selectPlans.map((opt,idx) => {
-								return (<option key={'field-opt-'+this.props.fieldId+"-"+idx} value={opt.id} data-placement="right" title={opt.name}>
-									 {(opt.name.length>20)?(string(opt.name).trim().substr(0, 15).concat("...").toString()):(opt.name)}
-								</option>);
-							})}
+						{this.props.fieldDef.selectPlans.map((opt, idx) => {
+							return (<option key={'field-opt-' + this.props.fieldId + "-" + idx} value={opt.id} data-placement="right" title={opt.name}>
+								{(opt.name.length > 20) ? (string(opt.name).trim().substr(0, 15).concat("...").toString()) : (opt.name)}
+							</option>);
+						})}
 					</select>
-				) 
+				)
 			} else {
 				strategicObjectivesPlans = (
 					<select
@@ -609,9 +659,9 @@ export default React.createClass({
 						id={this.props.fieldId}
 						ref={this.props.fieldId}
 						type={this.props.fieldDef.type}
-						>
+					>
 						<option key='field-opt-' data-placement="right" title={Messages.get("label.noGoalPlanRegistered")}>
-							{Messages.getEditable("label.noGoalPlanRegistered","fpdi-nav-label")}
+							{Messages.getEditable("label.noGoalPlanRegistered", "fpdi-nav-label")}
 						</option>
 					</select>
 				)
@@ -620,26 +670,26 @@ export default React.createClass({
 		return (!!this.props.undeletable || !!this.props.alterable ? (
 			<div className="panel panel-default panel-margins">
 				<div className="panel-heading attribute-input-opts">
-					{!this.props.vizualization ? 
-						<div className="edit-section-attribute"> 
-							<input defaultValue={this.props.fieldDef.label} className="edit-section-attribute-input" maxLength="255" ref="edit-input"/> <br/>
+					{!this.props.vizualization ?
+						<div className="edit-section-attribute">
+							<input defaultValue={this.props.fieldDef.label} className="edit-section-attribute-input" maxLength="255" ref="edit-input" /> <br />
 							<div className="formAlertError" ref="formAlertError-edit-input"></div>
 						</div>
-					:
+						:
 						<div>
 							<div className={(!!this.props.undeletable ? (this.state.menuHidden ? "" : "widthLimit pull-left") : "")}>
 								<b className="budget-title">{this.props.fieldDef.label}</b>
 							</div>
 							{strategicObjectivesPlans}
-							{(this.context.roles.MANAGER || _.contains(this.context.permissions, 
-									PermissionsTypes.MANAGE_DOCUMENT_PERMISSION)) && !this.context.planMacro.get("archived")?
+							{(this.context.roles.MANAGER || _.contains(this.context.permissions,
+								PermissionsTypes.MANAGE_DOCUMENT_PERMISSION)) && !this.context.planMacro.get("archived") ?
 								(!!this.props.undeletable ? <span type="submit" className="mdi mdi-delete attribute-input-edit inner"
-									title={Messages.get("label.deleteField")} onClick={this.delete}/> : "")
-							: ""}
-							<div className="clearfix"/>
+									title={Messages.get("label.deleteField")} onClick={this.delete} /> : "")
+								: ""}
+							<div className="clearfix" />
 						</div>
 					}
-					
+
 				</div>
 				<div>
 					{fieldEl}
@@ -647,19 +697,19 @@ export default React.createClass({
 				<div className="formAlertError" ref="formAlertError"></div>
 			</div>
 		) : (
-			<div className={"form-group form-group-sm" + (this.props.fieldDef.type == 'hidden' ? " hidden":"")}>
-				{this.props.vizualization && this.props.fieldDef.name == "indicator-type" ? "" : 
-				<label htmlFor={this.state.fieldId} className="fpdi-text-label">
-					{this.props.fieldDef.label}
-					{this.props.fieldDef.required && !this.props.vizualization ? <span className="fpdi-required">&nbsp;</span>:""}
-				</label>}
-				{fieldEl}
-				{typeof this.props.fieldDef.helpBox === 'string' ?
-					<p className="help-block">{this.props.fieldDef.helpBox}</p>
-					:this.props.fieldDef.helpBox
-				}
-				<div className="formAlertError" ref="formAlertError"></div>
-			</div>
-		));
+				<div className={"form-group form-group-sm" + (this.props.fieldDef.type == 'hidden' ? " hidden" : "")}>
+					{this.props.vizualization && this.props.fieldDef.name == "indicator-type" ? "" :
+						<label htmlFor={this.state.fieldId} className="fpdi-text-label">
+							{this.props.fieldDef.label}
+							{this.props.fieldDef.required && !this.props.vizualization ? <span className="fpdi-required">&nbsp;</span> : ""}
+						</label>}
+					{fieldEl}
+					{typeof this.props.fieldDef.helpBox === 'string' ?
+						<p className="help-block">{this.props.fieldDef.helpBox}</p>
+						: this.props.fieldDef.helpBox
+					}
+					<div className="formAlertError" ref="formAlertError"></div>
+				</div>
+			));
 	}
 });
